Reorganize managed ex-calib3d files into geometry/calib/stereo; update build-flag docs
Purely organizational follow-up to the calib3d split: drop the calib3d name
from managed files and group them by OpenCV 5 module (namespaces unchanged,
all flat OpenCvSharp).

- Modules/calib3d -> Modules/{geometry,calib,stereo} (StereoBM/SGBM/Matcher ->
  stereo; UsacParams -> geometry; enums distributed; folder removed).
- NativeMethods/calib3d -> split into NativeMethods/{geometry,calib,stereo}
  plus objdetect/NativeMethods_objdetect_chessboard.cs and
  imgproc/NativeMethods_imgproc_undistort.cs; fisheye -> calib, StereoMatcher
  -> stereo.
- Cv2_calib3d.cs -> Cv2_{geometry,calib,stereo}.cs (chessboard methods appended
  to Cv2_objdetect.cs, undistort/drawFrameAxes to Cv2_imgproc.cs);
  Cv2_calib3d.FishEye.cs -> Cv2_calib.FishEye.cs.
- docs/embedded-builds.md and build-opencvsharp-minimal-arm64.sh: NO_CALIB3D ->
  NO_GEOMETRY/NO_CALIB/NO_STEREO/NO_PTCLOUD, NO_FEATURES2D -> NO_FEATURES.

Verified: managed build (net8.0); Calib3d/Stereo/ImgProc/ObjDetect/PtCloud
tests pass (178).
