+++ src/OpenCvSharp/Cv2/Cv2_geometry.cs
@@ -1010,549 +1010,6 @@ static partial class Cv2
     }
 
     /// <summary>
-    /// initializes camera matrix from a few 3D points and the corresponding projections.
-    /// </summary>
-    /// <param name="objectPoints">Vector of vectors (vector&lt;vector&lt;Point3d&gt;&gt;) of the calibration pattern points in the calibration pattern coordinate space. In the old interface all the per-view vectors are concatenated.</param>
-    /// <param name="imagePoints">Vector of vectors (vector&lt;vector&lt;Point2d&gt;&gt;) of the projections of the calibration pattern points. In the old interface all the per-view vectors are concatenated.</param>
-    /// <param name="imageSize">Image size in pixels used to initialize the principal point.</param>
-    /// <param name="aspectRatio">If it is zero or negative, both f_x and f_y are estimated independently. Otherwise, f_x = f_y * aspectRatio .</param>
-    /// <returns></returns>
-    public static Mat InitCameraMatrix2D(
-        IEnumerable<Mat> objectPoints,
-        IEnumerable<Mat> imagePoints,
-        Size imageSize,
-        double aspectRatio = 1.0)
-    {
-        if (objectPoints is null)
-            throw new ArgumentNullException(nameof(objectPoints));
-        if (imagePoints is null)
-            throw new ArgumentNullException(nameof(imagePoints));
-
-        var objectPointsPtrs = objectPoints.Select(x => x.CvPtr).ToArray();
-        var imagePointsPtrs = imagePoints.Select(x => x.CvPtr).ToArray();
-
-        NativeMethods.HandleException(
-            NativeMethods.calib_initCameraMatrix2D_Mat(
-                objectPointsPtrs, objectPointsPtrs.Length,
-                imagePointsPtrs, imagePointsPtrs.Length, imageSize, aspectRatio,
-                out var matPtr));
-        return new Mat(matPtr);
-    }
-
-    /// <summary>
-    /// initializes camera matrix from a few 3D points and the corresponding projections.
-    /// </summary>
-    /// <param name="objectPoints">Vector of vectors of the calibration pattern points in the calibration pattern coordinate space. In the old interface all the per-view vectors are concatenated.</param>
-    /// <param name="imagePoints">Vector of vectors of the projections of the calibration pattern points. In the old interface all the per-view vectors are concatenated.</param>
-    /// <param name="imageSize">Image size in pixels used to initialize the principal point.</param>
-    /// <param name="aspectRatio">If it is zero or negative, both f_x and f_y are estimated independently. Otherwise, f_x = f_y * aspectRatio .</param>
-    /// <returns></returns>
-    public static Mat InitCameraMatrix2D(
-        IEnumerable<IEnumerable<Point3f>> objectPoints,
-        IEnumerable<IEnumerable<Point2f>> imagePoints,
-        Size imageSize,
-        double aspectRatio = 1.0)
-    {
-        if (objectPoints is null)
-            throw new ArgumentNullException(nameof(objectPoints));
-        if (imagePoints is null)
-            throw new ArgumentNullException(nameof(imagePoints));
-
-        using var opArray = new ArrayAddress2<Point3f>(objectPoints);
-        using var ipArray = new ArrayAddress2<Point2f>(imagePoints);
-        NativeMethods.HandleException(
-            NativeMethods.calib_initCameraMatrix2D_array(
-                opArray.GetPointer(), opArray.GetDim1Length(), opArray.GetDim2Lengths(),
-                ipArray.GetPointer(), ipArray.GetDim1Length(), ipArray.GetDim2Lengths(),
-                imageSize, aspectRatio, out var matPtr));
-        return new Mat(matPtr);
-    }
-
-    /// <summary>
-    /// Finds the positions of internal corners of the chessboard.
-    /// </summary>
-    /// <param name="image">Source chessboard view. It must be an 8-bit grayscale or color image.</param>
-    /// <param name="patternSize">Number of inner corners per a chessboard row and column 
-    /// ( patternSize = Size(points_per_row,points_per_colum) = Size(columns, rows) ).</param>
-    /// <param name="corners">Output array of detected corners.</param>
-    /// <param name="flags">Various operation flags that can be zero or a combination of the ChessboardFlag values</param>
-    /// <returns>The function returns true if all of the corners are found and they are placed in a certain order (row by row, left to right in every row). 
-    /// Otherwise, if the function fails to find all the corners or reorder them, it returns false.</returns>
-    public static bool FindChessboardCorners(
-        InputArray image,
-        Size patternSize,
-        OutputArray corners,
-        ChessboardFlags flags = ChessboardFlags.AdaptiveThresh | ChessboardFlags.NormalizeImage)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        if (corners is null)
-            throw new ArgumentNullException(nameof(corners));
-        image.ThrowIfDisposed();
-        corners.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.calib_findChessboardCorners_InputArray(
-                image.CvPtr, patternSize, corners.CvPtr, (int) flags, out var ret));
-        GC.KeepAlive(image);
-        corners.Fix();
-        return ret != 0;
-    }
-    /// <summary>
-    /// Finds the positions of internal corners of the chessboard.
-    /// </summary>
-    /// <param name="image">Source chessboard view. It must be an 8-bit grayscale or color image.</param>
-    /// <param name="patternSize">Number of inner corners per a chessboard row and column 
-    /// ( patternSize = Size(points_per_row,points_per_colum) = Size(columns, rows) ).</param>
-    /// <param name="corners">Output array of detected corners.</param>
-    /// <param name="flags">Various operation flags that can be zero or a combination of the ChessboardFlag values</param>
-    /// <returns>The function returns true if all of the corners are found and they are placed in a certain order (row by row, left to right in every row). 
-    /// Otherwise, if the function fails to find all the corners or reorder them, it returns false.</returns>
-    public static bool FindChessboardCorners(
-        InputArray image,
-        Size patternSize,
-        out Point2f[] corners,
-        ChessboardFlags flags = ChessboardFlags.AdaptiveThresh | ChessboardFlags.NormalizeImage)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        image.ThrowIfDisposed();
-
-        using var cornersVec = new VectorOfPoint2f();
-        NativeMethods.HandleException(
-            NativeMethods.calib_findChessboardCorners_vector(
-                image.CvPtr, patternSize, cornersVec.CvPtr, (int) flags, out var ret));
-        GC.KeepAlive(image);
-        corners = cornersVec.ToArray();
-        return ret != 0;
-    }
-
-    /// <summary>
-    /// Checks whether the image contains chessboard of the specific size or not.
-    /// </summary>
-    /// <param name="img"></param>
-    /// <param name="size"></param>
-    /// <returns></returns>
-    public static bool CheckChessboard(InputArray img, Size size)
-    {
-        if (img is null)
-            throw new ArgumentNullException(nameof(img));
-        img.ThrowIfDisposed();
-
-        NativeMethods.HandleException(
-            NativeMethods.objdetect_checkChessboard(img.CvPtr, size, out var ret));
-        GC.KeepAlive(img);
-        return ret != 0;
-    }
-
-    /// <summary>
-    /// Finds the positions of internal corners of the chessboard using a sector based approach.
-    /// </summary>
-    /// <param name="image">image Source chessboard view. It must be an 8-bit grayscale or color image.</param>
-    /// <param name="patternSize">Number of inner corners per a chessboard row and column
-    /// (patternSize = Size(points_per_row, points_per_column) = Size(columns, rows) ).</param>
-    /// <param name="corners">Output array of detected corners.</param>
-    /// <param name="flags">flags Various operation flags that can be zero or a combination of the ChessboardFlags values.</param>
-    /// <returns></returns>
-    public static bool FindChessboardCornersSB(
-        InputArray image, Size patternSize, OutputArray corners, ChessboardFlags flags = 0)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        if (corners is null)
-            throw new ArgumentNullException(nameof(corners));
-        image.ThrowIfDisposed();
-        corners.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.objdetect_findChessboardCornersSB_OutputArray(
-                image.CvPtr, patternSize, corners.CvPtr, (int) flags, out var ret));
-
-        GC.KeepAlive(image);
-        GC.KeepAlive(corners);
-        return ret != 0;
-    }
-
-    /// <summary>
-    /// Finds the positions of internal corners of the chessboard using a sector based approach.
-    /// </summary>
-    /// <param name="image">image Source chessboard view. It must be an 8-bit grayscale or color image.</param>
-    /// <param name="patternSize">Number of inner corners per a chessboard row and column
-    /// (patternSize = Size(points_per_row, points_per_column) = Size(columns, rows) ).</param>
-    /// <param name="corners">Output array of detected corners.</param>
-    /// <param name="flags">flags Various operation flags that can be zero or a combination of the ChessboardFlags values.</param>
-    /// <returns></returns>
-    public static bool FindChessboardCornersSB(
-        InputArray image, Size patternSize, out Point2f[] corners, ChessboardFlags flags = 0)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        image.ThrowIfDisposed();
-
-        using var cornersVec = new VectorOfPoint2f();
-        NativeMethods.HandleException(
-            NativeMethods.objdetect_findChessboardCornersSB_vector(
-                image.CvPtr, patternSize, cornersVec.CvPtr, (int) flags, out var ret));
-
-        corners = cornersVec.ToArray();
-        GC.KeepAlive(image);
-        return ret != 0;
-    }
-
-    /// <summary>
-    /// finds subpixel-accurate positions of the chessboard corners
-    /// </summary>
-    /// <param name="img"></param>
-    /// <param name="corners"></param>
-    /// <param name="regionSize"></param>
-    /// <returns></returns>
-    public static bool Find4QuadCornerSubpix(InputArray img, InputOutputArray corners, Size regionSize)
-    {
-        if (img is null)
-            throw new ArgumentNullException(nameof(img));
-        if (corners is null)
-            throw new ArgumentNullException(nameof(corners));
-        img.ThrowIfDisposed();
-        corners.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.objdetect_find4QuadCornerSubpix_InputArray(
-                img.CvPtr, corners.CvPtr, regionSize, out var ret));
-        GC.KeepAlive(img);
-        corners.Fix();
-        return ret != 0;
-    }
-    /// <summary>
-    /// finds subpixel-accurate positions of the chessboard corners
-    /// </summary>
-    /// <param name="img"></param>
-    /// <param name="corners"></param>
-    /// <param name="regionSize"></param>
-    /// <returns></returns>
-    public static bool Find4QuadCornerSubpix(InputArray img, Point2f[] corners, Size regionSize)
-    {
-        if (img is null)
-            throw new ArgumentNullException(nameof(img));
-        if (corners is null)
-            throw new ArgumentNullException(nameof(corners));
-        img.ThrowIfDisposed();
-
-        using var cornersVec = new VectorOfPoint2f(corners);
-        NativeMethods.HandleException(
-            NativeMethods.objdetect_find4QuadCornerSubpix_vector(
-                img.CvPtr, cornersVec.CvPtr, regionSize, out var ret));
-        GC.KeepAlive(img);
-
-        var newCorners = cornersVec.ToArray();
-        for (var i = 0; i < corners.Length; i++)
-        {
-            corners[i] = newCorners[i];
-        }
-
-        return ret != 0;
-    }
-
-    /// <summary>
-    /// Renders the detected chessboard corners.
-    /// </summary>
-    /// <param name="image">Destination image. It must be an 8-bit color image.</param>
-    /// <param name="patternSize">Number of inner corners per a chessboard row and column (patternSize = cv::Size(points_per_row,points_per_column)).</param>
-    /// <param name="corners">Array of detected corners, the output of findChessboardCorners.</param>
-    /// <param name="patternWasFound">Parameter indicating whether the complete board was found or not. The return value of findChessboardCorners() should be passed here.</param>
-    public static void DrawChessboardCorners(InputOutputArray image, Size patternSize,
-        InputArray corners, bool patternWasFound)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        if (corners is null)
-            throw new ArgumentNullException(nameof(corners));
-        image.ThrowIfNotReady();
-        corners.ThrowIfDisposed();
-
-        NativeMethods.HandleException(
-            NativeMethods.objdetect_drawChessboardCorners_InputArray(
-                image.CvPtr, patternSize, corners.CvPtr, patternWasFound ? 1 : 0));
-        GC.KeepAlive(corners);
-        image.Fix();
-    }
-
-    /// <summary>
-    /// Renders the detected chessboard corners.
-    /// </summary>
-    /// <param name="image">Destination image. It must be an 8-bit color image.</param>
-    /// <param name="patternSize">Number of inner corners per a chessboard row and column (patternSize = cv::Size(points_per_row,points_per_column)).</param>
-    /// <param name="corners">Array of detected corners, the output of findChessboardCorners.</param>
-    /// <param name="patternWasFound">Parameter indicating whether the complete board was found or not. The return value of findChessboardCorners() should be passed here.</param>
-    [SuppressMessage("Maintainability", "CA1508: Avoid dead conditional code")]
-    public static void DrawChessboardCorners(InputOutputArray image, Size patternSize,
-        IEnumerable<Point2f> corners, bool patternWasFound)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        if (corners is null)
-            throw new ArgumentNullException(nameof(corners));
-        image.ThrowIfNotReady();
-
-        var cornersArray = corners as Point2f[] ?? corners.ToArray();
-        NativeMethods.HandleException(
-            NativeMethods.objdetect_drawChessboardCorners_array(
-                image.CvPtr, patternSize, cornersArray, cornersArray.Length,
-                patternWasFound ? 1 : 0));
-        image.Fix();
-    }
-
-    /// <summary>
-    /// Draw axes of the world/object coordinate system from pose estimation.
-    /// </summary>
-    /// <param name="image">Input/output image. It must have 1 or 3 channels. The number of channels is not altered.</param>
-    /// <param name="cameraMatrix">Input 3x3 floating-point matrix of camera intrinsic parameters.</param>
-    /// <param name="distCoeffs">Input vector of distortion coefficients
-    /// \f$(k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6[, s_1, s_2, s_3, s_4[, \tau_x, \tau_y]]]])\f$ of
-    /// 4, 5, 8, 12 or 14 elements.If the vector is empty, the zero distortion coefficients are assumed.</param>
-    /// <param name="rvec">Rotation vector (see @ref Rodrigues ) that, together with tvec , brings points from
-    /// the model coordinate system to the camera coordinate system.</param>
-    /// <param name="tvec">Translation vector.</param>
-    /// <param name="length">Length of the painted axes in the same unit than tvec (usually in meters).</param>
-    /// <param name="thickness">Line thickness of the painted axes.</param>
-    /// <remarks>This function draws the axes of the world/object coordinate system w.r.t. to the camera frame.
-    /// OX is drawn in red, OY in green and OZ in blue.</remarks>
-    public static void DrawFrameAxes(
-        InputOutputArray image, InputArray cameraMatrix, InputArray distCoeffs,
-        InputArray rvec, InputArray tvec, float length, int thickness = 3)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        if (cameraMatrix is null)
-            throw new ArgumentNullException(nameof(cameraMatrix));
-        if (distCoeffs is null)
-            throw new ArgumentNullException(nameof(distCoeffs));
-        if (rvec is null)
-            throw new ArgumentNullException(nameof(rvec));
-        if (tvec is null)
-            throw new ArgumentNullException(nameof(tvec));
-        image.ThrowIfDisposed();
-        cameraMatrix.ThrowIfDisposed();
-        distCoeffs.ThrowIfDisposed();
-        rvec.ThrowIfDisposed();
-        tvec.ThrowIfDisposed();
-
-        NativeMethods.HandleException(
-            NativeMethods.imgproc_drawFrameAxes(
-                image.CvPtr, cameraMatrix.CvPtr, distCoeffs.CvPtr, rvec.CvPtr, tvec.CvPtr, length, thickness));
-
-        GC.KeepAlive(image);
-        GC.KeepAlive(cameraMatrix);
-        GC.KeepAlive(distCoeffs);
-        GC.KeepAlive(rvec);
-        GC.KeepAlive(tvec);
-    }
-
-    /// <summary>
-    /// Finds centers in the grid of circles.
-    /// </summary>
-    /// <param name="image">grid view of input circles; it must be an 8-bit grayscale or color image.</param>
-    /// <param name="patternSize">number of circles per row and column ( patternSize = Size(points_per_row, points_per_colum) ).</param>
-    /// <param name="centers">output array of detected centers.</param>
-    /// <param name="flags">various operation flags that can be one of the FindCirclesGridFlag values</param>
-    /// <param name="blobDetector">feature detector that finds blobs like dark circles on light background.</param>
-    /// <returns></returns>
-    public static bool FindCirclesGrid(
-        InputArray image,
-        Size patternSize,
-        OutputArray centers,
-        FindCirclesGridFlags flags = FindCirclesGridFlags.SymmetricGrid,
-        FeatureDetector? blobDetector = null)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        if (centers is null)
-            throw new ArgumentNullException(nameof(centers));
-        image.ThrowIfDisposed();
-        centers.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.calib_findCirclesGrid_InputArray(
-                image.CvPtr, patternSize, centers.CvPtr, (int) flags, ToPtr(blobDetector), out var ret));
-        GC.KeepAlive(image);
-        GC.KeepAlive(centers);
-        GC.KeepAlive(blobDetector);
-        centers.Fix();
-        return ret != 0;
-    }
-
-    /// <summary>
-    /// Finds centers in the grid of circles.
-    /// </summary>
-    /// <param name="image">grid view of input circles; it must be an 8-bit grayscale or color image.</param>
-    /// <param name="patternSize">number of circles per row and column ( patternSize = Size(points_per_row, points_per_colum) ).</param>
-    /// <param name="centers">output array of detected centers.</param>
-    /// <param name="flags">various operation flags that can be one of the FindCirclesGridFlag values</param>
-    /// <param name="blobDetector">feature detector that finds blobs like dark circles on light background.</param>
-    /// <returns></returns>
-    public static bool FindCirclesGrid(
-        InputArray image,
-        Size patternSize,
-        out Point2f[] centers,
-        FindCirclesGridFlags flags = FindCirclesGridFlags.SymmetricGrid,
-        FeatureDetector? blobDetector = null)
-    {
-        if (image is null)
-            throw new ArgumentNullException(nameof(image));
-        image.ThrowIfDisposed();
-
-        using var centersVec = new VectorOfPoint2f();
-        NativeMethods.HandleException(
-            NativeMethods.calib_findCirclesGrid_vector(
-                image.CvPtr, patternSize, centersVec.CvPtr, (int) flags, ToPtr(blobDetector), out var ret));
-        GC.KeepAlive(image);
-        GC.KeepAlive(blobDetector);
-        centers = centersVec.ToArray();
-        return ret != 0;
-    }
-
-    /// <summary>
-    /// finds intrinsic and extrinsic camera parameters from several fews of a known calibration pattern.
-    /// </summary>
-    /// <param name="objectPoints">In the new interface it is a vector of vectors of calibration pattern points in the calibration pattern coordinate space. 
-    /// The outer vector contains as many elements as the number of the pattern views. If the same calibration pattern is shown in each view and 
-    /// it is fully visible, all the vectors will be the same. Although, it is possible to use partially occluded patterns, or even different patterns 
-    /// in different views. Then, the vectors will be different. The points are 3D, but since they are in a pattern coordinate system, then, 
-    /// if the rig is planar, it may make sense to put the model to a XY coordinate plane so that Z-coordinate of each input object point is 0.
-    /// In the old interface all the vectors of object points from different views are concatenated together.</param>
-    /// <param name="imagePoints">In the new interface it is a vector of vectors of the projections of calibration pattern points. 
-    /// imagePoints.Count() and objectPoints.Count() and imagePoints[i].Count() must be equal to objectPoints[i].Count() for each i.</param>
-    /// <param name="imageSize">Size of the image used only to initialize the intrinsic camera matrix.</param>
-    /// <param name="cameraMatrix">Output 3x3 floating-point camera matrix. 
-    /// If CV_CALIB_USE_INTRINSIC_GUESS and/or CV_CALIB_FIX_ASPECT_RATIO are specified, some or all of fx, fy, cx, cy must be 
-    /// initialized before calling the function.</param>
-    /// <param name="distCoeffs">Output vector of distortion coefficients (k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6]]) of 4, 5, or 8 elements.</param>
-    /// <param name="rvecs">Output vector of rotation vectors (see Rodrigues() ) estimated for each pattern view. That is, each k-th rotation vector 
-    /// together with the corresponding k-th translation vector (see the next output parameter description) brings the calibration pattern 
-    /// from the model coordinate space (in which object points are specified) to the world coordinate space, that is, a real position of the 
-    /// calibration pattern in the k-th pattern view (k=0.. M -1)</param>
-    /// <param name="tvecs">Output vector of translation vectors estimated for each pattern view.</param>
-    /// <param name="flags">Different flags that may be zero or a combination of the CalibrationFlag values</param>
-    /// <param name="criteria">Termination criteria for the iterative optimization algorithm.</param>
-    /// <returns>Root mean square (RMS) re-projection error. A value below 1.0 is generally considered acceptable for a good calibration.</returns>
-    public static double CalibrateCamera(
-        IEnumerable<Mat> objectPoints,
-        IEnumerable<Mat> imagePoints,
-        Size imageSize,
-        InputOutputArray cameraMatrix,
-        InputOutputArray distCoeffs,
-        out Mat[] rvecs,
-        out Mat[] tvecs,
-        CalibrationFlags flags = CalibrationFlags.None,
-        TermCriteria? criteria = null)
-    {
-        if (objectPoints is null)
-            throw new ArgumentNullException(nameof(objectPoints));
-        if (imagePoints is null)
-            throw new ArgumentNullException(nameof(imagePoints));
-        if (cameraMatrix is null)
-            throw new ArgumentNullException(nameof(cameraMatrix));
-        if (distCoeffs is null)
-            throw new ArgumentNullException(nameof(distCoeffs));
-        cameraMatrix.ThrowIfNotReady();
-        distCoeffs.ThrowIfNotReady();
-
-        var criteria0 = criteria.GetValueOrDefault(
-            new TermCriteria(CriteriaTypes.Count | CriteriaTypes.Eps, 30, Double.Epsilon));
-
-        var objectPointsPtrs = objectPoints.Select(x => x.CvPtr).ToArray();
-        var imagePointsPtrs = imagePoints.Select(x => x.CvPtr).ToArray();
-
-        using var rvecsVec = new VectorOfMat();
-        using var tvecsVec = new VectorOfMat();
-        NativeMethods.HandleException(
-            NativeMethods.calib_calibrateCamera_InputArray(
-                objectPointsPtrs, objectPointsPtrs.Length,
-                imagePointsPtrs, objectPointsPtrs.Length,
-                imageSize, cameraMatrix.CvPtr, distCoeffs.CvPtr,
-                rvecsVec.CvPtr, tvecsVec.CvPtr, (int) flags, criteria0, out var ret));
-        GC.KeepAlive(cameraMatrix);
-        GC.KeepAlive(distCoeffs);
-        GC.KeepAlive(objectPoints);
-        GC.KeepAlive(imagePoints);
-        rvecs = rvecsVec.ToArray();
-        tvecs = tvecsVec.ToArray();
-
-        cameraMatrix.Fix();
-        distCoeffs.Fix();
-        return ret;
-    }
-
-    /// <summary>
-    /// finds intrinsic and extrinsic camera parameters from several fews of a known calibration pattern.
-    /// </summary>
-    /// <param name="objectPoints">In the new interface it is a vector of vectors of calibration pattern points in the calibration pattern coordinate space. 
-    /// The outer vector contains as many elements as the number of the pattern views. If the same calibration pattern is shown in each view and 
-    /// it is fully visible, all the vectors will be the same. Although, it is possible to use partially occluded patterns, or even different patterns 
-    /// in different views. Then, the vectors will be different. The points are 3D, but since they are in a pattern coordinate system, then, 
-    /// if the rig is planar, it may make sense to put the model to a XY coordinate plane so that Z-coordinate of each input object point is 0.
-    /// In the old interface all the vectors of object points from different views are concatenated together.</param>
-    /// <param name="imagePoints">In the new interface it is a vector of vectors of the projections of calibration pattern points. 
-    /// imagePoints.Count() and objectPoints.Count() and imagePoints[i].Count() must be equal to objectPoints[i].Count() for each i.</param>
-    /// <param name="imageSize">Size of the image used only to initialize the intrinsic camera matrix.</param>
-    /// <param name="cameraMatrix">Output 3x3 floating-point camera matrix. 
-    /// If CV_CALIB_USE_INTRINSIC_GUESS and/or CV_CALIB_FIX_ASPECT_RATIO are specified, some or all of fx, fy, cx, cy must be 
-    /// initialized before calling the function.</param>
-    /// <param name="distCoeffs">Output vector of distortion coefficients (k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6]]) of 4, 5, or 8 elements.</param>
-    /// <param name="rvecs">Output vector of rotation vectors (see Rodrigues() ) estimated for each pattern view. That is, each k-th rotation vector 
-    /// together with the corresponding k-th translation vector (see the next output parameter description) brings the calibration pattern 
-    /// from the model coordinate space (in which object points are specified) to the world coordinate space, that is, a real position of the 
-    /// calibration pattern in the k-th pattern view (k=0.. M -1)</param>
-    /// <param name="tvecs">Output vector of translation vectors estimated for each pattern view.</param>
-    /// <param name="flags">Different flags that may be zero or a combination of the CalibrationFlag values</param>
-    /// <param name="criteria">Termination criteria for the iterative optimization algorithm.</param>
-    /// <returns>Root mean square (RMS) re-projection error. A value below 1.0 is generally considered acceptable for a good calibration.</returns>
-    public static double CalibrateCamera(
-        IEnumerable<IEnumerable<Point3f>> objectPoints,
-        IEnumerable<IEnumerable<Point2f>> imagePoints,
-        Size imageSize,
-        double[,] cameraMatrix,
-        double[] distCoeffs,
-        out Vec3d[] rvecs,
-        out Vec3d[] tvecs,
-        CalibrationFlags flags = CalibrationFlags.None,
-        TermCriteria? criteria = null)
-    {
-        if (objectPoints is null)
-            throw new ArgumentNullException(nameof(objectPoints));
-        if (imagePoints is null)
-            throw new ArgumentNullException(nameof(imagePoints));
-        if (cameraMatrix is null)
-            throw new ArgumentNullException(nameof(cameraMatrix));
-        if (distCoeffs is null)
-            throw new ArgumentNullException(nameof(distCoeffs));
-
-        var criteria0 = criteria.GetValueOrDefault(
-            new TermCriteria(CriteriaTypes.Count | CriteriaTypes.Eps, 30, Double.Epsilon));
-
-        using var op = new ArrayAddress2<Point3f>(objectPoints);
-        using var ip = new ArrayAddress2<Point2f>(imagePoints);
-        using var rvecsVec = new VectorOfMat();
-        using var tvecsVec = new VectorOfMat();
-        unsafe
-        {
-            fixed (double* cameraMatrixPtr = cameraMatrix)
-            {
-                NativeMethods.HandleException(
-                    NativeMethods.calib_calibrateCamera_vector(
-                        op.GetPointer(), op.GetDim1Length(), op.GetDim2Lengths(),
-                        ip.GetPointer(), ip.GetDim1Length(), ip.GetDim2Lengths(),
-                        imageSize, cameraMatrixPtr, distCoeffs, distCoeffs.Length,
-                        rvecsVec.CvPtr, tvecsVec.CvPtr, (int) flags, criteria0, out var ret));
-                var rvecsM = rvecsVec.ToArray();
-                var tvecsM = tvecsVec.ToArray();
-                rvecs = rvecsM.Select(m => m.Get<Vec3d>(0)).ToArray();
-                tvecs = tvecsM.Select(m => m.Get<Vec3d>(0)).ToArray();
-                return ret;
-            }
-        }
-    }
-
-    /// <summary>
     /// computes several useful camera characteristics from the camera matrix, camera frame resolution and the physical sensor size.
     /// </summary>
     /// <param name="cameraMatrix">Input camera matrix that can be estimated by calibrateCamera() or stereoCalibrate() .</param>
@@ -1615,761 +1072,6 @@ static partial class Cv2
                         out principalPoint, out aspectRatio));
             }
         }
-    }
-
-    /// <summary>
-    /// finds intrinsic and extrinsic parameters of a stereo camera
-    /// </summary>
-    /// <param name="objectPoints">Vector of vectors of the calibration pattern points.</param>
-    /// <param name="imagePoints1">Vector of vectors of the projections of the calibration pattern points, observed by the first camera.</param>
-    /// <param name="imagePoints2">Vector of vectors of the projections of the calibration pattern points, observed by the second camera.</param>
-    /// <param name="cameraMatrix1">Input/output first camera matrix</param>
-    /// <param name="distCoeffs1">Input/output vector of distortion coefficients (k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6]]) of 4, 5, or 8 elements. 
-    /// The output vector length depends on the flags.</param>
-    /// <param name="cameraMatrix2"> Input/output second camera matrix. The parameter is similar to cameraMatrix1 .</param>
-    /// <param name="distCoeffs2">Input/output lens distortion coefficients for the second camera. The parameter is similar to distCoeffs1 .</param>
-    /// <param name="imageSize">Size of the image used only to initialize intrinsic camera matrix.</param>
-    /// <param name="R">Output rotation matrix between the 1st and the 2nd camera coordinate systems.</param>
-    /// <param name="T">Output translation vector between the coordinate systems of the cameras.</param>
-    /// <param name="E">Output essential matrix.</param>
-    /// <param name="F">Output fundamental matrix.</param>
-    /// <param name="criteria">Termination criteria for the iterative optimization algorithm.</param>
-    /// <param name="flags">Different flags that may be zero or a combination of the CalibrationFlag values</param>
-    /// <returns></returns>
-    public static double StereoCalibrate(
-        IEnumerable<InputArray> objectPoints,
-        IEnumerable<InputArray> imagePoints1,
-        IEnumerable<InputArray> imagePoints2,
-        InputOutputArray cameraMatrix1, InputOutputArray distCoeffs1,
-        InputOutputArray cameraMatrix2, InputOutputArray distCoeffs2,
-        Size imageSize, OutputArray R,
-        OutputArray T, OutputArray E, OutputArray F,
-        CalibrationFlags flags = CalibrationFlags.FixIntrinsic,
-        TermCriteria? criteria = null)
-    {
-        if (objectPoints is null)
-            throw new ArgumentNullException(nameof(objectPoints));
-        if (imagePoints1 is null)
-            throw new ArgumentNullException(nameof(imagePoints1));
-        if (imagePoints2 is null)
-            throw new ArgumentNullException(nameof(imagePoints2));
-        if (cameraMatrix1 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix1));
-        if (distCoeffs1 is null)
-            throw new ArgumentNullException(nameof(distCoeffs1));
-        if (cameraMatrix2 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix2));
-        if (distCoeffs2 is null)
-            throw new ArgumentNullException(nameof(distCoeffs2));
-        cameraMatrix1.ThrowIfDisposed();
-        distCoeffs1.ThrowIfDisposed();
-        cameraMatrix2.ThrowIfDisposed();
-        distCoeffs2.ThrowIfDisposed();
-        cameraMatrix1.ThrowIfNotReady();
-        cameraMatrix2.ThrowIfNotReady();
-        distCoeffs1.ThrowIfNotReady();
-        distCoeffs2.ThrowIfNotReady();
-
-        var opPtrs = objectPoints.Select(x => x.CvPtr).ToArray();
-        var ip1Ptrs = imagePoints1.Select(x => x.CvPtr).ToArray();
-        var ip2Ptrs = imagePoints2.Select(x => x.CvPtr).ToArray();
-
-        var criteria0 = criteria.GetValueOrDefault(
-            new TermCriteria(CriteriaTypes.Count | CriteriaTypes.Eps, 30, 1e-6));
-
-        NativeMethods.HandleException(
-            NativeMethods.calib_stereoCalibrate_InputArray(
-                opPtrs, opPtrs.Length,
-                ip1Ptrs, ip1Ptrs.Length, ip2Ptrs, ip2Ptrs.Length,
-                cameraMatrix1.CvPtr, distCoeffs1.CvPtr,
-                cameraMatrix2.CvPtr, distCoeffs2.CvPtr,
-                imageSize, ToPtr(R), ToPtr(T), ToPtr(E), ToPtr(F),
-                (int) flags, criteria0, out var ret));
-
-        GC.KeepAlive(cameraMatrix1);
-        GC.KeepAlive(distCoeffs1);
-        GC.KeepAlive(cameraMatrix2);
-        GC.KeepAlive(distCoeffs2);
-        GC.KeepAlive(R);
-        GC.KeepAlive(T);
-        GC.KeepAlive(E);
-        GC.KeepAlive(F);
-        GC.KeepAlive(objectPoints);
-        GC.KeepAlive(imagePoints1);
-        GC.KeepAlive(imagePoints2);
-        cameraMatrix1.Fix();
-        distCoeffs1.Fix();
-        cameraMatrix2.Fix();
-        distCoeffs2.Fix();
-        R.Fix();
-        T.Fix();
-        E.Fix();
-        F.Fix();
-
-        return ret;
-    }
-
-    /// <summary>
-    /// finds intrinsic and extrinsic parameters of a stereo camera
-    /// </summary>
-    /// <param name="objectPoints">Vector of vectors of the calibration pattern points.</param>
-    /// <param name="imagePoints1">Vector of vectors of the projections of the calibration pattern points, observed by the first camera.</param>
-    /// <param name="imagePoints2">Vector of vectors of the projections of the calibration pattern points, observed by the second camera.</param>
-    /// <param name="cameraMatrix1">Input/output first camera matrix</param>
-    /// <param name="distCoeffs1">Input/output vector of distortion coefficients (k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6]]) of 4, 5, or 8 elements. 
-    /// The output vector length depends on the flags.</param>
-    /// <param name="cameraMatrix2"> Input/output second camera matrix. The parameter is similar to cameraMatrix1 .</param>
-    /// <param name="distCoeffs2">Input/output lens distortion coefficients for the second camera. The parameter is similar to distCoeffs1 .</param>
-    /// <param name="imageSize">Size of the image used only to initialize intrinsic camera matrix.</param>
-    /// <param name="R">Output rotation matrix between the 1st and the 2nd camera coordinate systems.</param>
-    /// <param name="T">Output translation vector between the coordinate systems of the cameras.</param>
-    /// <param name="E">Output essential matrix.</param>
-    /// <param name="F">Output fundamental matrix.</param>
-    /// <param name="criteria">Termination criteria for the iterative optimization algorithm.</param>
-    /// <param name="flags">Different flags that may be zero or a combination of the CalibrationFlag values</param>
-    /// <returns></returns>
-    public static double StereoCalibrate(
-        IEnumerable<Mat> objectPoints,
-        IEnumerable<Mat> imagePoints1,
-        IEnumerable<Mat> imagePoints2,
-        Mat cameraMatrix1, Mat distCoeffs1,
-        Mat cameraMatrix2, Mat distCoeffs2,
-        Size imageSize, Mat R, Mat T, Mat E, Mat F,
-        CalibrationFlags flags = CalibrationFlags.FixIntrinsic,
-        TermCriteria? criteria = null)
-    {
-        if (objectPoints is null)
-            throw new ArgumentNullException(nameof(objectPoints));
-        if (imagePoints1 is null)
-            throw new ArgumentNullException(nameof(imagePoints1));
-        if (imagePoints2 is null)
-            throw new ArgumentNullException(nameof(imagePoints2));
-        if (cameraMatrix1 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix1));
-        if (distCoeffs1 is null)
-            throw new ArgumentNullException(nameof(distCoeffs1));
-        if (cameraMatrix2 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix2));
-        if (distCoeffs2 is null)
-            throw new ArgumentNullException(nameof(distCoeffs2));
-        if (R is null)
-            throw new ArgumentNullException(nameof(R));
-        if (T is null)
-            throw new ArgumentNullException(nameof(T));
-        if (E is null)
-            throw new ArgumentNullException(nameof(E));
-        if (F is null)
-            throw new ArgumentNullException(nameof(F));
-
-        var opPtrs = objectPoints.Select(x => x.CvPtr).ToArray();
-        var ip1Ptrs = imagePoints1.Select(x => x.CvPtr).ToArray();
-        var ip2Ptrs = imagePoints2.Select(x => x.CvPtr).ToArray();
-
-        var criteria0 = criteria.GetValueOrDefault(
-            new TermCriteria(CriteriaTypes.Count | CriteriaTypes.Eps, 30, 1e-6));
-
-        NativeMethods.HandleException(
-            NativeMethods.calib_stereoCalibrate_Mat(
-                opPtrs, opPtrs.Length,
-                ip1Ptrs, ip1Ptrs.Length,
-                ip2Ptrs, ip2Ptrs.Length,
-                cameraMatrix1.CvPtr, distCoeffs1.CvPtr,
-                cameraMatrix2.CvPtr, distCoeffs2.CvPtr,
-                imageSize, R.CvPtr, T.CvPtr, E.CvPtr, F.CvPtr,
-                (int)flags, criteria0, out var ret));
-
-        GC.KeepAlive(objectPoints);
-        GC.KeepAlive(imagePoints1);
-        GC.KeepAlive(imagePoints2);
-        GC.KeepAlive(cameraMatrix1);
-        GC.KeepAlive(distCoeffs1);
-        GC.KeepAlive(cameraMatrix2);
-        GC.KeepAlive(distCoeffs2);
-        GC.KeepAlive(R);
-        GC.KeepAlive(T);
-        GC.KeepAlive(E);
-        GC.KeepAlive(F);
-
-        return ret;
-    }
-
-    /// <summary>
-    /// finds intrinsic and extrinsic parameters of a stereo camera
-    /// </summary>
-    /// <param name="objectPoints">Vector of vectors of the calibration pattern points.</param>
-    /// <param name="imagePoints1">Vector of vectors of the projections of the calibration pattern points, observed by the first camera.</param>
-    /// <param name="imagePoints2">Vector of vectors of the projections of the calibration pattern points, observed by the second camera.</param>
-    /// <param name="cameraMatrix1">Input/output first camera matrix</param>
-    /// <param name="distCoeffs1">Input/output vector of distortion coefficients (k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6]]) of 4, 5, or 8 elements. 
-    /// The output vector length depends on the flags.</param>
-    /// <param name="cameraMatrix2"> Input/output second camera matrix. The parameter is similar to cameraMatrix1 .</param>
-    /// <param name="distCoeffs2">Input/output lens distortion coefficients for the second camera. The parameter is similar to distCoeffs1 .</param>
-    /// <param name="imageSize">Size of the image used only to initialize intrinsic camera matrix.</param>
-    /// <param name="R">Output rotation matrix between the 1st and the 2nd camera coordinate systems.</param>
-    /// <param name="T">Output translation vector between the coordinate systems of the cameras.</param>
-    /// <param name="E">Output essential matrix.</param>
-    /// <param name="F">Output fundamental matrix.</param>
-    /// <param name="criteria">Termination criteria for the iterative optimization algorithm.</param>
-    /// <param name="flags">Different flags that may be zero or a combination of the CalibrationFlag values</param>
-    /// <returns></returns>
-    public static double StereoCalibrate(
-        IEnumerable<IEnumerable<Point3f>> objectPoints,
-        IEnumerable<IEnumerable<Point2f>> imagePoints1,
-        IEnumerable<IEnumerable<Point2f>> imagePoints2,
-        double[,] cameraMatrix1, double[] distCoeffs1,
-        double[,] cameraMatrix2, double[] distCoeffs2,
-        Size imageSize, OutputArray R,
-        OutputArray T, OutputArray E, OutputArray F,
-        CalibrationFlags flags = CalibrationFlags.FixIntrinsic,
-        TermCriteria? criteria = null)
-    {
-        if (objectPoints is null)
-            throw new ArgumentNullException(nameof(objectPoints));
-        if (imagePoints1 is null)
-            throw new ArgumentNullException(nameof(imagePoints1));
-        if (imagePoints2 is null)
-            throw new ArgumentNullException(nameof(imagePoints2));
-        if (cameraMatrix1 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix1));
-        if (distCoeffs1 is null)
-            throw new ArgumentNullException(nameof(distCoeffs1));
-        if (cameraMatrix2 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix2));
-        if (distCoeffs2 is null)
-            throw new ArgumentNullException(nameof(distCoeffs2));
-
-        var criteria0 = criteria.GetValueOrDefault(
-            new TermCriteria(CriteriaTypes.Count | CriteriaTypes.Eps, 30, 1e-6));
-
-        using var op = new ArrayAddress2<Point3f>(objectPoints);
-        using var ip1 = new ArrayAddress2<Point2f>(imagePoints1);
-        using var ip2 = new ArrayAddress2<Point2f>(imagePoints2);
-        unsafe
-        {
-            fixed (double* cameraMatrix1Ptr = cameraMatrix1)
-            fixed (double* cameraMatrix2Ptr = cameraMatrix2)
-            {
-                NativeMethods.HandleException(
-                    NativeMethods.calib_stereoCalibrate_array(
-                        op.GetPointer(), op.GetDim1Length(), op.GetDim2Lengths(),
-                        ip1.GetPointer(), ip1.GetDim1Length(), ip1.GetDim2Lengths(),
-                        ip2.GetPointer(), ip2.GetDim1Length(), ip2.GetDim2Lengths(),
-                        cameraMatrix1Ptr, distCoeffs1, distCoeffs1.Length,
-                        cameraMatrix2Ptr, distCoeffs2, distCoeffs2.Length,
-                        imageSize, ToPtr(R), ToPtr(T), ToPtr(E), ToPtr(F),
-                        (int) flags, criteria0, out var ret));
-                GC.KeepAlive(R);
-                GC.KeepAlive(T);
-                GC.KeepAlive(E);
-                GC.KeepAlive(F);
-                return ret;
-            }
-        }
-    }
-
-    /// <summary>
-    /// computes the rectification transformation for a stereo camera from its intrinsic and extrinsic parameters
-    /// </summary>
-    /// <param name="cameraMatrix1">First camera matrix.</param>
-    /// <param name="distCoeffs1">First camera distortion parameters.</param>
-    /// <param name="cameraMatrix2">Second camera matrix.</param>
-    /// <param name="distCoeffs2">Second camera distortion parameters.</param>
-    /// <param name="imageSize">Size of the image used for stereo calibration.</param>
-    /// <param name="R">Rotation matrix between the coordinate systems of the first and the second cameras.</param>
-    /// <param name="T">Translation vector between coordinate systems of the cameras.</param>
-    /// <param name="R1">Output 3x3 rectification transform (rotation matrix) for the first camera.</param>
-    /// <param name="R2"> Output 3x3 rectification transform (rotation matrix) for the second camera.</param>
-    /// <param name="P1">Output 3x4 projection matrix in the new (rectified) coordinate systems for the first camera.</param>
-    /// <param name="P2">Output 3x4 projection matrix in the new (rectified) coordinate systems for the second camera.</param>
-    /// <param name="Q">Output 4x4 disparity-to-depth mapping matrix (see reprojectImageTo3D() ).</param>
-    /// <param name="flags">Operation flags that may be zero or CV_CALIB_ZERO_DISPARITY. 
-    /// If the flag is set, the function makes the principal points of each camera have the same pixel coordinates in the rectified views. 
-    /// And if the flag is not set, the function may still shift the images in the horizontal or vertical direction (depending on the orientation of epipolar lines) to maximize the useful image area.</param>
-    /// <param name="alpha">Free scaling parameter. 
-    /// If it is -1 or absent, the function performs the default scaling. Otherwise, the parameter should be between 0 and 1. 
-    /// alpha=0 means that the rectified images are zoomed and shifted so that only valid pixels are visible (no black areas after rectification). 
-    /// alpha=1 means that the rectified image is decimated and shifted so that all the pixels from the original images from the cameras are retained 
-    /// in the rectified images (no source image pixels are lost). Obviously, any intermediate value yields an intermediate result between those two extreme cases.</param>
-    /// <param name="newImageSize">New image resolution after rectification. The same size should be passed to initUndistortRectifyMap(). When (0,0) is passed (default), it is set to the original imageSize . 
-    /// Setting it to larger value can help you preserve details in the original image, especially when there is a big radial distortion.</param>
-    public static void StereoRectify(InputArray cameraMatrix1, InputArray distCoeffs1,
-        InputArray cameraMatrix2, InputArray distCoeffs2,
-        Size imageSize, InputArray R, InputArray T,
-        OutputArray R1, OutputArray R2,
-        OutputArray P1, OutputArray P2,
-        OutputArray Q,
-        StereoRectificationFlags flags = StereoRectificationFlags.ZeroDisparity,
-        double alpha = -1, Size? newImageSize = null)
-    {
-        var newImageSize0 = newImageSize.GetValueOrDefault(new Size(0, 0));
-        StereoRectify(cameraMatrix1, distCoeffs1, cameraMatrix2, distCoeffs2,
-            imageSize, R, T, R1, R2, P1, P2, Q, flags, alpha, newImageSize0,
-            out _, out _);
-    }
-
-    /// <summary>
-    /// computes the rectification transformation for a stereo camera from its intrinsic and extrinsic parameters
-    /// </summary>
-    /// <param name="cameraMatrix1">First camera matrix.</param>
-    /// <param name="distCoeffs1">First camera distortion parameters.</param>
-    /// <param name="cameraMatrix2">Second camera matrix.</param>
-    /// <param name="distCoeffs2">Second camera distortion parameters.</param>
-    /// <param name="imageSize">Size of the image used for stereo calibration.</param>
-    /// <param name="R">Rotation matrix between the coordinate systems of the first and the second cameras.</param>
-    /// <param name="T">Translation vector between coordinate systems of the cameras.</param>
-    /// <param name="R1">Output 3x3 rectification transform (rotation matrix) for the first camera.</param>
-    /// <param name="R2"> Output 3x3 rectification transform (rotation matrix) for the second camera.</param>
-    /// <param name="P1">Output 3x4 projection matrix in the new (rectified) coordinate systems for the first camera.</param>
-    /// <param name="P2">Output 3x4 projection matrix in the new (rectified) coordinate systems for the second camera.</param>
-    /// <param name="Q">Output 4x4 disparity-to-depth mapping matrix (see reprojectImageTo3D() ).</param>
-    /// <param name="flags">Operation flags that may be zero or CV_CALIB_ZERO_DISPARITY. 
-    /// If the flag is set, the function makes the principal points of each camera have the same pixel coordinates in the rectified views. 
-    /// And if the flag is not set, the function may still shift the images in the horizontal or vertical direction (depending on the orientation of epipolar lines) to maximize the useful image area.</param>
-    /// <param name="alpha">Free scaling parameter. 
-    /// If it is -1 or absent, the function performs the default scaling. Otherwise, the parameter should be between 0 and 1. 
-    /// alpha=0 means that the rectified images are zoomed and shifted so that only valid pixels are visible (no black areas after rectification). 
-    /// alpha=1 means that the rectified image is decimated and shifted so that all the pixels from the original images from the cameras are retained 
-    /// in the rectified images (no source image pixels are lost). Obviously, any intermediate value yields an intermediate result between those two extreme cases.</param>
-    /// <param name="newImageSize">New image resolution after rectification. The same size should be passed to initUndistortRectifyMap(). When (0,0) is passed (default), it is set to the original imageSize . 
-    /// Setting it to larger value can help you preserve details in the original image, especially when there is a big radial distortion.</param>
-    /// <param name="validPixROI1">Optional output rectangles inside the rectified images where all the pixels are valid. If alpha=0 , the ROIs cover the whole images. 
-    /// Otherwise, they are likely to be smaller.</param>
-    /// <param name="validPixROI2">Optional output rectangles inside the rectified images where all the pixels are valid. If alpha=0 , the ROIs cover the whole images. 
-    /// Otherwise, they are likely to be smaller.</param>
-    public static void StereoRectify(InputArray cameraMatrix1, InputArray distCoeffs1,
-        InputArray cameraMatrix2, InputArray distCoeffs2,
-        Size imageSize, InputArray R, InputArray T,
-        OutputArray R1, OutputArray R2,
-        OutputArray P1, OutputArray P2,
-        OutputArray Q, StereoRectificationFlags flags,
-        double alpha, Size newImageSize,
-        out Rect validPixROI1, out Rect validPixROI2)
-    {
-        if (cameraMatrix1 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix1));
-        if (distCoeffs1 is null)
-            throw new ArgumentNullException(nameof(distCoeffs1));
-        if (cameraMatrix2 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix2));
-        if (distCoeffs2 is null)
-            throw new ArgumentNullException(nameof(distCoeffs2));
-        if (R is null)
-            throw new ArgumentNullException(nameof(R));
-        if (T is null)
-            throw new ArgumentNullException(nameof(T));
-        if (R1 is null)
-            throw new ArgumentNullException(nameof(R1));
-        if (R2 is null)
-            throw new ArgumentNullException(nameof(R2));
-        if (P1 is null)
-            throw new ArgumentNullException(nameof(P1));
-        if (P2 is null)
-            throw new ArgumentNullException(nameof(P2));
-        if (Q is null)
-            throw new ArgumentNullException(nameof(Q));
-        cameraMatrix1.ThrowIfDisposed();
-        distCoeffs1.ThrowIfDisposed();
-        cameraMatrix2.ThrowIfDisposed();
-        distCoeffs2.ThrowIfDisposed();
-        R.ThrowIfDisposed();
-        T.ThrowIfDisposed();
-        R1.ThrowIfNotReady();
-        R2.ThrowIfNotReady();
-        P1.ThrowIfNotReady();
-        P2.ThrowIfNotReady();
-        Q.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.stereo_stereoRectify_InputArray(
-                cameraMatrix1.CvPtr, distCoeffs1.CvPtr,
-                cameraMatrix2.CvPtr, distCoeffs2.CvPtr,
-                imageSize, R.CvPtr, T.CvPtr,
-                R1.CvPtr, R2.CvPtr, P1.CvPtr, P2.CvPtr, Q.CvPtr,
-                (int) flags, alpha, newImageSize, out validPixROI1, out validPixROI2));
-
-        GC.KeepAlive(cameraMatrix1);
-        GC.KeepAlive(distCoeffs1);
-        GC.KeepAlive(cameraMatrix2);
-        GC.KeepAlive(distCoeffs2);
-        GC.KeepAlive(R);
-        GC.KeepAlive(T);
-        GC.KeepAlive(R1);
-        GC.KeepAlive(R2);
-        GC.KeepAlive(P1);
-        GC.KeepAlive(P2);
-        GC.KeepAlive(Q);
-
-        R1.Fix();
-        R2.Fix();
-        P1.Fix();
-        P2.Fix();
-        Q.Fix();
-    }
-
-    /// <summary>
-    /// computes the rectification transformation for a stereo camera from its intrinsic and extrinsic parameters
-    /// </summary>
-    /// <param name="cameraMatrix1">First camera matrix.</param>
-    /// <param name="distCoeffs1">First camera distortion parameters.</param>
-    /// <param name="cameraMatrix2">Second camera matrix.</param>
-    /// <param name="distCoeffs2">Second camera distortion parameters.</param>
-    /// <param name="imageSize">Size of the image used for stereo calibration.</param>
-    /// <param name="R">Rotation matrix between the coordinate systems of the first and the second cameras.</param>
-    /// <param name="T">Translation vector between coordinate systems of the cameras.</param>
-    /// <param name="R1">Output 3x3 rectification transform (rotation matrix) for the first camera.</param>
-    /// <param name="R2"> Output 3x3 rectification transform (rotation matrix) for the second camera.</param>
-    /// <param name="P1">Output 3x4 projection matrix in the new (rectified) coordinate systems for the first camera.</param>
-    /// <param name="P2">Output 3x4 projection matrix in the new (rectified) coordinate systems for the second camera.</param>
-    /// <param name="Q">Output 4x4 disparity-to-depth mapping matrix (see reprojectImageTo3D() ).</param>
-    /// <param name="flags">Operation flags that may be zero or CV_CALIB_ZERO_DISPARITY. 
-    /// If the flag is set, the function makes the principal points of each camera have the same pixel coordinates in the rectified views. 
-    /// And if the flag is not set, the function may still shift the images in the horizontal or vertical direction (depending on the orientation of epipolar lines) to maximize the useful image area.</param>
-    /// <param name="alpha">Free scaling parameter. 
-    /// If it is -1 or absent, the function performs the default scaling. Otherwise, the parameter should be between 0 and 1. 
-    /// alpha=0 means that the rectified images are zoomed and shifted so that only valid pixels are visible (no black areas after rectification). 
-    /// alpha=1 means that the rectified image is decimated and shifted so that all the pixels from the original images from the cameras are retained 
-    /// in the rectified images (no source image pixels are lost). Obviously, any intermediate value yields an intermediate result between those two extreme cases.</param>
-    /// <param name="newImageSize">New image resolution after rectification. The same size should be passed to initUndistortRectifyMap(). When (0,0) is passed (default), it is set to the original imageSize . 
-    /// Setting it to larger value can help you preserve details in the original image, especially when there is a big radial distortion.</param>
-    public static void StereoRectify(double[,] cameraMatrix1, double[] distCoeffs1,
-        double[,] cameraMatrix2, double[] distCoeffs2,
-        Size imageSize, double[,] R, double[] T,
-        out double[,] R1, out double[,] R2,
-        out double[,] P1, out double[,] P2,
-        out double[,] Q,
-        StereoRectificationFlags flags = StereoRectificationFlags.ZeroDisparity,
-        double alpha = -1, Size? newImageSize = null)
-    {
-        var newImageSize0 = newImageSize.GetValueOrDefault(new Size(0, 0));
-        StereoRectify(
-            cameraMatrix1, distCoeffs1,
-            cameraMatrix2, distCoeffs2,
-            imageSize, R, T,
-            out R1, out R2, out P1, out P2, out Q,
-            flags, alpha, newImageSize0, out _, out _);
-    }
-
-    /// <summary>
-    /// computes the rectification transformation for a stereo camera from its intrinsic and extrinsic parameters
-    /// </summary>
-    /// <param name="cameraMatrix1">First camera matrix.</param>
-    /// <param name="distCoeffs1">First camera distortion parameters.</param>
-    /// <param name="cameraMatrix2">Second camera matrix.</param>
-    /// <param name="distCoeffs2">Second camera distortion parameters.</param>
-    /// <param name="imageSize">Size of the image used for stereo calibration.</param>
-    /// <param name="R">Rotation matrix between the coordinate systems of the first and the second cameras.</param>
-    /// <param name="T">Translation vector between coordinate systems of the cameras.</param>
-    /// <param name="R1">Output 3x3 rectification transform (rotation matrix) for the first camera.</param>
-    /// <param name="R2"> Output 3x3 rectification transform (rotation matrix) for the second camera.</param>
-    /// <param name="P1">Output 3x4 projection matrix in the new (rectified) coordinate systems for the first camera.</param>
-    /// <param name="P2">Output 3x4 projection matrix in the new (rectified) coordinate systems for the second camera.</param>
-    /// <param name="Q">Output 4x4 disparity-to-depth mapping matrix (see reprojectImageTo3D() ).</param>
-    /// <param name="flags">Operation flags that may be zero or CV_CALIB_ZERO_DISPARITY. 
-    /// If the flag is set, the function makes the principal points of each camera have the same pixel coordinates in the rectified views. 
-    /// And if the flag is not set, the function may still shift the images in the horizontal or vertical direction (depending on the orientation of epipolar lines) to maximize the useful image area.</param>
-    /// <param name="alpha">Free scaling parameter. 
-    /// If it is -1 or absent, the function performs the default scaling. Otherwise, the parameter should be between 0 and 1. 
-    /// alpha=0 means that the rectified images are zoomed and shifted so that only valid pixels are visible (no black areas after rectification). 
-    /// alpha=1 means that the rectified image is decimated and shifted so that all the pixels from the original images from the cameras are retained 
-    /// in the rectified images (no source image pixels are lost). Obviously, any intermediate value yields an intermediate result between those two extreme cases.</param>
-    /// <param name="newImageSize">New image resolution after rectification. The same size should be passed to initUndistortRectifyMap(). When (0,0) is passed (default), it is set to the original imageSize . 
-    /// Setting it to larger value can help you preserve details in the original image, especially when there is a big radial distortion.</param>
-    /// <param name="validPixROI1">Optional output rectangles inside the rectified images where all the pixels are valid. If alpha=0 , the ROIs cover the whole images. 
-    /// Otherwise, they are likely to be smaller.</param>
-    /// <param name="validPixROI2">Optional output rectangles inside the rectified images where all the pixels are valid. If alpha=0 , the ROIs cover the whole images. 
-    /// Otherwise, they are likely to be smaller.</param>
-    public static void StereoRectify(double[,] cameraMatrix1, double[] distCoeffs1,
-        double[,] cameraMatrix2, double[] distCoeffs2,
-        Size imageSize, double[,] R, double[] T,
-        out double[,] R1, out double[,] R2,
-        out double[,] P1, out double[,] P2,
-        out double[,] Q, StereoRectificationFlags flags,
-        double alpha, Size newImageSize,
-        out Rect validPixROI1, out Rect validPixROI2)
-    {
-        if (cameraMatrix1 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix1));
-        if (distCoeffs1 is null)
-            throw new ArgumentNullException(nameof(distCoeffs1));
-        if (cameraMatrix2 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix2));
-        if (distCoeffs2 is null)
-            throw new ArgumentNullException(nameof(distCoeffs2));
-        if (R is null)
-            throw new ArgumentNullException(nameof(R));
-        if (T is null)
-            throw new ArgumentNullException(nameof(T));
-
-        R1 = new double[3, 3];
-        R2 = new double[3, 3];
-        P1 = new double[3, 4];
-        P2 = new double[3, 4];
-        Q = new double[4, 4];
-
-        unsafe
-        {
-            fixed (double* cameraMatrix1Ptr = cameraMatrix1)
-            fixed (double* cameraMatrix2Ptr = cameraMatrix2)
-            fixed (double* RPtr = R)
-            fixed (double* R1Ptr = R1)
-            fixed (double* R2Ptr = R2)
-            fixed (double* P1Ptr = P1)
-            fixed (double* P2Ptr = P2)
-            fixed (double* QPtr = Q)
-            {
-                NativeMethods.HandleException(
-                    NativeMethods.stereo_stereoRectify_array(
-                        cameraMatrix1Ptr, distCoeffs1, distCoeffs1.Length,
-                        cameraMatrix2Ptr, distCoeffs2, distCoeffs2.Length,
-                        imageSize, RPtr, T,
-                        R1Ptr, R2Ptr, P1Ptr, P2Ptr, QPtr,
-                        (int) flags, alpha, newImageSize, out validPixROI1, out validPixROI2));
-            }
-        }
-    }
-
-    /// <summary>
-    /// computes the rectification transformation for an uncalibrated stereo camera (zero distortion is assumed)
-    /// </summary>
-    /// <param name="points1">Array of feature points in the first image.</param>
-    /// <param name="points2">The corresponding points in the second image. 
-    /// The same formats as in findFundamentalMat() are supported.</param>
-    /// <param name="F">Input fundamental matrix. It can be computed from the same set 
-    /// of point pairs using findFundamentalMat() .</param>
-    /// <param name="imgSize">Size of the image.</param>
-    /// <param name="H1">Output rectification homography matrix for the first image.</param>
-    /// <param name="H2">Output rectification homography matrix for the second image.</param>
-    /// <param name="threshold">Optional threshold used to filter out the outliers. 
-    /// If the parameter is greater than zero, all the point pairs that do not comply 
-    /// with the epipolar geometry (that is, the points for which |points2[i]^T * F * points1[i]| > threshold ) 
-    /// are rejected prior to computing the homographies. Otherwise, all the points are considered inliers.</param>
-    /// <returns></returns>
-    public static bool StereoRectifyUncalibrated(InputArray points1, InputArray points2,
-        InputArray F, Size imgSize,
-        OutputArray H1, OutputArray H2,
-        double threshold = 5)
-    {
-        if (points1 is null)
-            throw new ArgumentNullException(nameof(points1));
-        if (points2 is null)
-            throw new ArgumentNullException(nameof(points2));
-        if (F is null)
-            throw new ArgumentNullException(nameof(F));
-        if (H1 is null)
-            throw new ArgumentNullException(nameof(H1));
-        if (H2 is null)
-            throw new ArgumentNullException(nameof(H2));
-        points1.ThrowIfDisposed();
-        points2.ThrowIfDisposed();
-        F.ThrowIfDisposed();
-        H1.ThrowIfNotReady();
-        H2.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.stereo_stereoRectifyUncalibrated_InputArray(
-                points1.CvPtr, points2.CvPtr, F.CvPtr, imgSize, H1.CvPtr, H2.CvPtr, threshold, out var ret));
-
-        GC.KeepAlive(points1);
-        GC.KeepAlive(points2);
-        GC.KeepAlive(F);
-        GC.KeepAlive(H1);
-        GC.KeepAlive(H2);
-        H1.Fix();
-        H2.Fix();
-        return ret != 0;
-    }
-
-    /// <summary>
-    /// computes the rectification transformation for an uncalibrated stereo camera (zero distortion is assumed)
-    /// </summary>
-    /// <param name="points1">Array of feature points in the first image.</param>
-    /// <param name="points2">The corresponding points in the second image. 
-    /// The same formats as in findFundamentalMat() are supported.</param>
-    /// <param name="F">Input fundamental matrix. It can be computed from the same set 
-    /// of point pairs using findFundamentalMat() .</param>
-    /// <param name="imgSize">Size of the image.</param>
-    /// <param name="H1">Output rectification homography matrix for the first image.</param>
-    /// <param name="H2">Output rectification homography matrix for the second image.</param>
-    /// <param name="threshold">Optional threshold used to filter out the outliers. 
-    /// If the parameter is greater than zero, all the point pairs that do not comply 
-    /// with the epipolar geometry (that is, the points for which |points2[i]^T * F * points1[i]| > threshold ) 
-    /// are rejected prior to computing the homographies. Otherwise, all the points are considered inliers.</param>
-    /// <returns></returns>
-    [SuppressMessage("Maintainability", "CA1508: Avoid dead conditional code")]
-    public static bool StereoRectifyUncalibrated(
-        IEnumerable<Point2d> points1,
-        IEnumerable<Point2d> points2,
-        double[,] F, Size imgSize,
-        out double[,] H1, out double[,] H2,
-        double threshold = 5)
-    {
-        if (points1 is null)
-            throw new ArgumentNullException(nameof(points1));
-        if (points2 is null)
-            throw new ArgumentNullException(nameof(points2));
-        if (F is null)
-            throw new ArgumentNullException(nameof(F));
-        if (F.GetLength(0) != 3 || F.GetLength(1) != 3)
-            throw new ArgumentException("F != double[3,3]");
-
-        var points1Array = points1 as Point2d[] ?? points1.ToArray();
-        var points2Array = points2 as Point2d[] ?? points2.ToArray();
-
-        H1 = new double[3, 3];
-        H2 = new double[3, 3];
-
-        unsafe
-        {
-            fixed (double* FPtr = F)
-            fixed (double* H1Ptr = H1)
-            fixed (double* H2Ptr = H1)
-            {
-                NativeMethods.HandleException(
-                    NativeMethods.stereo_stereoRectifyUncalibrated_array(
-                        points1Array, points1Array.Length,
-                        points2Array, points2Array.Length,
-                        FPtr, imgSize, H1Ptr, H2Ptr, threshold, out var ret));
-                return ret != 0;
-            }
-        }
-    }
-
-    /// <summary>
-    /// computes the rectification transformations for 3-head camera, where all the heads are on the same line.
-    /// </summary>
-    /// <param name="cameraMatrix1"></param>
-    /// <param name="distCoeffs1"></param>
-    /// <param name="cameraMatrix2"></param>
-    /// <param name="distCoeffs2"></param>
-    /// <param name="cameraMatrix3"></param>
-    /// <param name="distCoeffs3"></param>
-    /// <param name="imgpt1"></param>
-    /// <param name="imgpt3"></param>
-    /// <param name="imageSize"></param>
-    /// <param name="R12"></param>
-    /// <param name="T12"></param>
-    /// <param name="R13"></param>
-    /// <param name="T13"></param>
-    /// <param name="R1"></param>
-    /// <param name="R2"></param>
-    /// <param name="R3"></param>
-    /// <param name="P1"></param>
-    /// <param name="P2"></param>
-    /// <param name="P3"></param>
-    /// <param name="Q"></param>
-    /// <param name="alpha"></param>
-    /// <param name="newImgSize"></param>
-    /// <param name="roi1"></param>
-    /// <param name="roi2"></param>
-    /// <param name="flags"></param>
-    /// <returns></returns>
-    public static float Rectify3Collinear(
-        InputArray cameraMatrix1, InputArray distCoeffs1,
-        InputArray cameraMatrix2, InputArray distCoeffs2,
-        InputArray cameraMatrix3, InputArray distCoeffs3,
-        IEnumerable<InputArray> imgpt1, IEnumerable<InputArray> imgpt3,
-        Size imageSize, InputArray R12, InputArray T12,
-        InputArray R13, InputArray T13,
-        OutputArray R1, OutputArray R2, OutputArray R3,
-        OutputArray P1, OutputArray P2, OutputArray P3,
-        OutputArray Q, double alpha, Size newImgSize,
-        out Rect roi1, out Rect roi2, StereoRectificationFlags flags)
-    {
-        if (cameraMatrix1 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix1));
-        if (distCoeffs1 is null)
-            throw new ArgumentNullException(nameof(distCoeffs1));
-        if (cameraMatrix2 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix2));
-        if (distCoeffs2 is null)
-            throw new ArgumentNullException(nameof(distCoeffs2));
-        if (cameraMatrix3 is null)
-            throw new ArgumentNullException(nameof(cameraMatrix3));
-        if (distCoeffs3 is null)
-            throw new ArgumentNullException(nameof(distCoeffs3));
-        if (imgpt1 is null)
-            throw new ArgumentNullException(nameof(imgpt1));
-        if (imgpt3 is null)
-            throw new ArgumentNullException(nameof(imgpt3));
-        if (R12 is null)
-            throw new ArgumentNullException(nameof(R12));
-        if (T12 is null)
-            throw new ArgumentNullException(nameof(T12));
-        if (R13 is null)
-            throw new ArgumentNullException(nameof(R13));
-        if (T13 is null)
-            throw new ArgumentNullException(nameof(T13));
-        if (R1 is null)
-            throw new ArgumentNullException(nameof(R1));
-        if (R2 is null)
-            throw new ArgumentNullException(nameof(R2));
-        if (R3 is null)
-            throw new ArgumentNullException(nameof(R3));
-        if (P1 is null)
-            throw new ArgumentNullException(nameof(P1));
-        if (P2 is null)
-            throw new ArgumentNullException(nameof(P2));
-        if (P3 is null)
-            throw new ArgumentNullException(nameof(P3));
-        if (Q is null)
-            throw new ArgumentNullException(nameof(Q));
-        cameraMatrix1.ThrowIfDisposed();
-        distCoeffs1.ThrowIfDisposed();
-        cameraMatrix2.ThrowIfDisposed();
-        distCoeffs2.ThrowIfDisposed();
-        cameraMatrix3.ThrowIfDisposed();
-        distCoeffs3.ThrowIfDisposed();
-        R12.ThrowIfDisposed();
-        T12.ThrowIfDisposed();
-        R13.ThrowIfDisposed();
-        T13.ThrowIfDisposed();
-        R1.ThrowIfNotReady();
-        R2.ThrowIfNotReady();
-        R3.ThrowIfNotReady();
-        P1.ThrowIfNotReady();
-        P2.ThrowIfNotReady();
-        P3.ThrowIfNotReady();
-        Q.ThrowIfNotReady();
-
-        var imgpt1Ptrs = imgpt1.Select(x => x.CvPtr).ToArray();
-        var imgpt3Ptrs = imgpt3.Select(x => x.CvPtr).ToArray();
-        NativeMethods.HandleException(
-            NativeMethods.stereo_rectify3Collinear_InputArray(
-                cameraMatrix1.CvPtr, distCoeffs1.CvPtr,
-                cameraMatrix2.CvPtr, distCoeffs2.CvPtr,
-                cameraMatrix3.CvPtr, distCoeffs3.CvPtr,
-                imgpt1Ptrs, imgpt1Ptrs.Length, imgpt3Ptrs, imgpt3Ptrs.Length,
-                imageSize, R12.CvPtr, T12.CvPtr, R13.CvPtr, T13.CvPtr,
-                R1.CvPtr, R2.CvPtr, R3.CvPtr, P1.CvPtr, P2.CvPtr, P3.CvPtr,
-                Q.CvPtr, alpha, newImgSize, out roi1, out roi2, (int) flags, out var ret));
-
-        GC.KeepAlive(cameraMatrix1);
-        GC.KeepAlive(distCoeffs1);
-        GC.KeepAlive(cameraMatrix2);
-        GC.KeepAlive(distCoeffs2);
-        GC.KeepAlive(cameraMatrix3);
-        GC.KeepAlive(distCoeffs3);
-        GC.KeepAlive(imgpt1);
-        GC.KeepAlive(imgpt3);
-        GC.KeepAlive(R12);
-        GC.KeepAlive(T12);
-        GC.KeepAlive(R13);
-        GC.KeepAlive(T13);
-        GC.KeepAlive(R1);
-        GC.KeepAlive(R2);
-        GC.KeepAlive(R3);
-        GC.KeepAlive(P1);
-        GC.KeepAlive(P2);
-        GC.KeepAlive(P3);
-        GC.KeepAlive(Q);
-        R1.Fix();
-        R2.Fix();
-        R3.Fix();
-        P1.Fix();
-        P2.Fix();
-        P3.Fix();
-        Q.Fix();
-        return ret;
     }
 
     /// <summary>
@@ -2451,281 +1153,6 @@ static partial class Cv2
 
         using var mat = Mat<double>.FromNativePointer(matPtr);
         return mat.ToRectangularArray();
-    }
-
-    /// <summary>
-    /// Computes Hand-Eye calibration.
-    /// 
-    /// The function performs the Hand-Eye calibration using various methods. One approach consists in estimating the
-    /// rotation then the translation(separable solutions) and the following methods are implemented:
-    /// - R.Tsai, R.Lenz A New Technique for Fully Autonomous and Efficient 3D Robotics Hand/EyeCalibration \cite Tsai89
-    /// - F.Park, B.Martin Robot Sensor Calibration: Solving AX = XB on the Euclidean Group \cite Park94
-    /// - R.Horaud, F.Dornaika Hand-Eye Calibration \cite Horaud95
-    ///
-    /// Another approach consists in estimating simultaneously the rotation and the translation(simultaneous solutions),
-    /// with the following implemented method:
-    /// - N.Andreff, R.Horaud, B.Espiau On-line Hand-Eye Calibration \cite Andreff99
-    /// - K.Daniilidis Hand-Eye Calibration Using Dual Quaternions \cite Daniilidis98
-    /// </summary>
-    /// <param name="R_gripper2base">Rotation part extracted from the homogeneous matrix that
-    /// transforms a pointexpressed in the gripper frame to the robot base frame that contains the rotation
-    /// matrices for all the transformationsfrom gripper frame to robot base frame.</param>
-    /// <param name="t_gripper2base">Translation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the gripper frame to the robot base frame.
-    /// This is a vector(`vector&lt;Mat&gt;`) that contains the translation vectors for all the transformations
-    /// from gripper frame to robot base frame.</param>
-    /// <param name="R_target2cam">Rotation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the target frame to the camera frame.
-    /// This is a vector(`vector&lt;Mat&gt;`) that contains the rotation matrices for all the transformations
-    /// from calibration target frame to camera frame.</param>
-    /// <param name="t_target2cam">Rotation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the target frame to the camera frame.
-    /// This is a vector(`vector&lt;Mat&gt;`) that contains the translation vectors for all the transformations
-    /// from calibration target frame to camera frame.</param>
-    /// <param name="R_cam2gripper">Estimated rotation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the camera frame to the gripper frame.</param>
-    /// <param name="t_cam2gripper">Estimated translation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the camera frame to the gripper frame.</param>
-    /// <param name="method">One of the implemented Hand-Eye calibration method</param>
-    [SuppressMessage("Maintainability", "CA1508: Avoid dead conditional code")]
-    public static void CalibrateHandEye(
-        IEnumerable<Mat> R_gripper2base,
-        IEnumerable<Mat> t_gripper2base,
-        IEnumerable<Mat> R_target2cam,
-        IEnumerable<Mat> t_target2cam,
-        OutputArray R_cam2gripper,
-        OutputArray t_cam2gripper,
-        HandEyeCalibrationMethod method = HandEyeCalibrationMethod.TSAI)
-    {
-        if (R_gripper2base is null)
-            throw new ArgumentNullException(nameof(R_gripper2base));
-        if (t_gripper2base is null)
-            throw new ArgumentNullException(nameof(t_gripper2base));
-        if (R_target2cam is null)
-            throw new ArgumentNullException(nameof(R_target2cam));
-        if (t_target2cam is null)
-            throw new ArgumentNullException(nameof(t_target2cam));
-        if (R_cam2gripper is null)
-            throw new ArgumentNullException(nameof(R_cam2gripper));
-        if (t_cam2gripper is null)
-            throw new ArgumentNullException(nameof(t_cam2gripper));
-        R_cam2gripper.ThrowIfNotReady();
-        t_cam2gripper.ThrowIfNotReady();
-
-        var R_gripper2baseArray = R_gripper2base as Mat[] ?? R_gripper2base.ToArray();
-        var t_gripper2baseArray = t_gripper2base as Mat[] ?? t_gripper2base.ToArray();
-        var R_target2camArray = R_target2cam as Mat[] ?? R_target2cam.ToArray();
-        var t_target2camArray = t_target2cam as Mat[] ?? t_target2cam.ToArray();
-        if (R_gripper2baseArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(R_gripper2base));
-        if (t_gripper2baseArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(t_gripper2base));
-        if (R_target2camArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(R_target2cam));
-        if (t_target2camArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(t_target2cam));
-
-        var R_gripper2basePtrArray = R_gripper2baseArray.Select(m => m.CvPtr).ToArray();
-        var t_gripper2basePtrArray = t_gripper2baseArray.Select(m => m.CvPtr).ToArray();
-        var R_target2camPtrArray = R_target2camArray.Select(m => m.CvPtr).ToArray();
-        var t_target2camPtrArray = t_target2camArray.Select(m => m.CvPtr).ToArray();
-        NativeMethods.HandleException(
-            NativeMethods.calib_calibrateHandEye(
-                R_gripper2basePtrArray, R_gripper2basePtrArray.Length,
-                t_gripper2basePtrArray, t_gripper2basePtrArray.Length,
-                R_target2camPtrArray, R_target2camPtrArray.Length,
-                t_target2camPtrArray, t_target2camPtrArray.Length,
-                R_cam2gripper.CvPtr, t_cam2gripper.CvPtr, (int)method));
-
-        GC.KeepAlive(R_gripper2base);
-        GC.KeepAlive(t_gripper2base);
-        GC.KeepAlive(R_target2cam);
-        GC.KeepAlive(t_target2cam);
-        R_cam2gripper.Fix();
-        t_cam2gripper.Fix();
-
-        foreach (var mat in R_gripper2baseArray) GC.KeepAlive(mat);
-        foreach (var mat in t_gripper2baseArray) GC.KeepAlive(mat);
-        foreach (var mat in R_target2camArray) GC.KeepAlive(mat);
-        foreach (var mat in t_target2camArray) GC.KeepAlive(mat);
-    }
-
-    /// <summary>
-    /// Computes Robot-World/Hand-Eye calibration.
-    /// The function performs the Robot-World/Hand-Eye calibration using various methods. One approach consists in estimating the
-    /// rotation then the translation(separable solutions):
-    /// - M.Shah, Solving the robot-world/hand-eye calibration problem using the kronecker product \cite Shah2013SolvingTR
-    /// </summary>
-    /// <param name="R_world2cam">[in] R_world2cam Rotation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the world frame to the camera frame. This is a vector of Mat that contains the rotation,
-    /// `(3x3)` rotation matrices or `(3x1)` rotation vectors,for all the transformations from world frame to the camera frame.</param>
-    /// <param name="t_world2cam">[in] Translation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the world frame to the camera frame. This is a vector (`vector&lt;Mat&gt;`) that contains the `(3x1)`
-    /// translation vectors for all the transformations from world frame to the camera frame.</param>
-    /// <param name="R_base2gripper">[in] Rotation part extracted from the homogeneous matrix that transforms a point expressed
-    /// in the robot base frame to the gripper frame. This is a vector (`vector&lt;Mat&gt;`) that contains the rotation,
-    /// `(3x3)` rotation matrices or `(3x1)` rotation vectors, for all the transformations from robot base frame to the gripper frame.</param>
-    /// <param name="t_base2gripper">[in] Rotation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the robot base frame to the gripper frame. This is a vector (`vector&lt;Mat&gt;`) that contains the
-    /// `(3x1)` translation vectors for all the transformations from robot base frame to the gripper frame.</param>
-    /// <param name="R_base2world">[out] R_base2world Estimated `(3x3)` rotation part extracted from the homogeneous matrix
-    /// that transforms a point expressed in the robot base frame to the world frame.</param>
-    /// <param name="t_base2world">[out] t_base2world Estimated `(3x1)` translation part extracted from the homogeneous matrix
-    /// that transforms a point expressed in the robot base frame to the world frame.</param>
-    /// <param name="R_gripper2cam">[out] R_gripper2cam Estimated `(3x3)` rotation part extracted from the homogeneous matrix
-    /// that transforms a point expressed in the gripper frame to the camera frame.</param>
-    /// <param name="t_gripper2cam">[out] Estimated `(3x1)` translation part extracted from the homogeneous matrix that
-    /// transforms a pointexpressed in the gripper frame to the camera frame.</param>
-    /// <param name="method">One of the implemented Robot-World/Hand-Eye calibration method</param>
-    [SuppressMessage("Maintainability", "CA1508: Avoid dead conditional code")]
-    public static void CalibrateRobotWorldHandEye(
-        IEnumerable<Mat> R_world2cam,
-        IEnumerable<Mat> t_world2cam,
-        IEnumerable<Mat> R_base2gripper,
-        IEnumerable<Mat> t_base2gripper,
-        OutputArray R_base2world, 
-        OutputArray t_base2world,
-        OutputArray R_gripper2cam,
-        OutputArray t_gripper2cam,
-        RobotWorldHandEyeCalibrationMethod method = RobotWorldHandEyeCalibrationMethod.SHAH)
-    {
-        if (R_world2cam is null)
-            throw new ArgumentNullException(nameof(R_world2cam));
-        if (t_world2cam is null)
-            throw new ArgumentNullException(nameof(t_world2cam));
-        if (R_base2gripper is null)
-            throw new ArgumentNullException(nameof(R_base2gripper));
-        if (t_base2gripper is null)
-            throw new ArgumentNullException(nameof(t_base2gripper));
-        if (R_base2world is null)
-            throw new ArgumentNullException(nameof(R_base2world));
-        if (t_base2world is null)
-            throw new ArgumentNullException(nameof(t_base2world));
-        if (R_gripper2cam is null)
-            throw new ArgumentNullException(nameof(R_gripper2cam));
-        if (t_gripper2cam is null)
-            throw new ArgumentNullException(nameof(t_gripper2cam));
-        R_base2world.ThrowIfNotReady();
-        t_base2world.ThrowIfNotReady();
-        R_gripper2cam.ThrowIfNotReady();
-        t_gripper2cam.ThrowIfNotReady();
-        var R_world2camArray = R_world2cam as Mat[] ?? R_world2cam.ToArray();
-        var t_world2camArray = t_world2cam as Mat[] ?? t_world2cam.ToArray();
-        var R_base2gripperArray = R_base2gripper as Mat[] ?? R_base2gripper.ToArray();
-        var t_base2gripperArray = t_base2gripper as Mat[] ?? t_base2gripper.ToArray();
-        if (R_world2camArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(R_world2cam));
-        if (t_world2camArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(t_world2cam));
-        if (R_base2gripperArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(R_base2gripper));
-        if (t_base2gripperArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(t_base2gripper));
-
-        var R_base2worldPtrArray = R_world2camArray.Select(m => m.CvPtr).ToArray();
-        var t_world2camPtrArray = t_world2camArray.Select(m => m.CvPtr).ToArray();
-        var R_base2gripperPtrArray = R_base2gripperArray.Select(m => m.CvPtr).ToArray();
-        var t_base2gripperPtrArray = t_base2gripperArray.Select(m => m.CvPtr).ToArray();
-        NativeMethods.HandleException(
-            NativeMethods.calib_calibrateRobotWorldHandEye_OutputArray(
-                R_base2worldPtrArray, R_base2worldPtrArray.Length,
-                t_world2camPtrArray, t_world2camPtrArray.Length,
-                R_base2gripperPtrArray, R_base2gripperPtrArray.Length,
-                t_base2gripperPtrArray, t_base2gripperPtrArray.Length,
-                R_base2world.CvPtr, t_base2world.CvPtr, R_gripper2cam.CvPtr, t_gripper2cam.CvPtr,
-                (int)method));
-
-        R_base2world.Fix();
-        t_base2world.Fix();
-        R_gripper2cam.Fix();
-        t_gripper2cam.Fix();
-        foreach (var mat in R_world2camArray) GC.KeepAlive(mat);
-        foreach (var mat in t_world2camArray) GC.KeepAlive(mat);
-        foreach (var mat in R_base2gripperArray) GC.KeepAlive(mat);
-        foreach (var mat in t_base2gripperArray) GC.KeepAlive(mat);
-    }
-
-    /// <summary>
-    /// omputes Robot-World/Hand-Eye calibration.
-    /// The function performs the Robot-World/Hand-Eye calibration using various methods. One approach consists in estimating the
-    /// rotation then the translation(separable solutions):
-    /// - M.Shah, Solving the robot-world/hand-eye calibration problem using the kronecker product \cite Shah2013SolvingTR
-    /// </summary>
-    /// <param name="R_world2cam">[in] R_world2cam Rotation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the world frame to the camera frame. This is a vector of Mat that contains the rotation,
-    /// `(3x3)` rotation matrices or `(3x1)` rotation vectors,for all the transformations from world frame to the camera frame.</param>
-    /// <param name="t_world2cam">[in] Translation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the world frame to the camera frame. This is a vector (`vector&lt;Mat&gt;`) that contains the `(3x1)`
-    /// translation vectors for all the transformations from world frame to the camera frame.</param>
-    /// <param name="R_base2gripper">[in] Rotation part extracted from the homogeneous matrix that transforms a point expressed
-    /// in the robot base frame to the gripper frame. This is a vector (`vector&lt;Mat&gt;`) that contains the rotation,
-    /// `(3x3)` rotation matrices or `(3x1)` rotation vectors, for all the transformations from robot base frame to the gripper frame.</param>
-    /// <param name="t_base2gripper">[in] Rotation part extracted from the homogeneous matrix that transforms a point
-    /// expressed in the robot base frame to the gripper frame. This is a vector (`vector&lt;Mat&gt;`) that contains the
-    /// `(3x1)` translation vectors for all the transformations from robot base frame to the gripper frame.</param>
-    /// <param name="R_base2world">[out] R_base2world Estimated `(3x3)` rotation part extracted from the homogeneous matrix
-    /// that transforms a point expressed in the robot base frame to the world frame.</param>
-    /// <param name="t_base2world">[out] t_base2world Estimated `(3x1)` translation part extracted from the homogeneous matrix
-    /// that transforms a point expressed in the robot base frame to the world frame.</param>
-    /// <param name="R_gripper2cam">[out] R_gripper2cam Estimated `(3x3)` rotation part extracted from the homogeneous matrix
-    /// that transforms a point expressed in the gripper frame to the camera frame.</param>
-    /// <param name="t_gripper2cam">[out] Estimated `(3x1)` translation part extracted from the homogeneous matrix that
-    /// transforms a pointexpressed in the gripper frame to the camera frame.</param>
-    /// <param name="method">One of the implemented Robot-World/Hand-Eye calibration method</param>
-    [SuppressMessage("Maintainability", "CA1508: Avoid dead conditional code")]
-    public static void CalibrateRobotWorldHandEye(
-        IEnumerable<Mat> R_world2cam,
-        IEnumerable<Mat> t_world2cam,
-        IEnumerable<Mat> R_base2gripper,
-        IEnumerable<Mat> t_base2gripper,
-        out double[,] R_base2world,
-        out double[] t_base2world,
-        out double[,] R_gripper2cam,
-        out double[] t_gripper2cam,
-        RobotWorldHandEyeCalibrationMethod method = RobotWorldHandEyeCalibrationMethod.SHAH)
-    {
-        if (R_world2cam is null)
-            throw new ArgumentNullException(nameof(R_world2cam));
-        if (t_world2cam is null)
-            throw new ArgumentNullException(nameof(t_world2cam));
-        if (R_base2gripper is null)
-            throw new ArgumentNullException(nameof(R_base2gripper));
-        if (t_base2gripper is null)
-            throw new ArgumentNullException(nameof(t_base2gripper));
-        var R_world2camArray = R_world2cam as Mat[] ?? R_world2cam.ToArray();
-        var t_world2camArray = t_world2cam as Mat[] ?? t_world2cam.ToArray();
-        var R_base2gripperArray = R_base2gripper as Mat[] ?? R_base2gripper.ToArray();
-        var t_base2gripperArray = t_base2gripper as Mat[] ?? t_base2gripper.ToArray();
-        if (R_world2camArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(R_world2cam));
-        if (t_world2camArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(t_world2cam));
-        if (R_base2gripperArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(R_base2gripper));
-        if (t_base2gripperArray.Length == 0)
-            throw new ArgumentException("Empty sequence", nameof(t_base2gripper));
-
-        var R_base2worldPtrArray = R_world2camArray.Select(m => m.CvPtr).ToArray();
-        var t_world2camPtrArray = t_world2camArray.Select(m => m.CvPtr).ToArray();
-        var R_base2gripperPtrArray = R_base2gripperArray.Select(m => m.CvPtr).ToArray();
-        var t_base2gripperPtrArray = t_base2gripperArray.Select(m => m.CvPtr).ToArray();
-        R_base2world = new double[3, 3];
-        t_base2world = new double[3];
-        R_gripper2cam = new double[3, 3];
-        t_gripper2cam = new double[3];
-        NativeMethods.HandleException(
-            NativeMethods.calib_calibrateRobotWorldHandEye_Pointer(
-                R_base2worldPtrArray, R_base2worldPtrArray.Length,
-                t_world2camPtrArray, t_world2camPtrArray.Length,
-                R_base2gripperPtrArray, R_base2gripperPtrArray.Length,
-                t_base2gripperPtrArray, t_base2gripperPtrArray.Length,
-                R_base2world, t_base2world, R_gripper2cam, t_gripper2cam,
-                (int) method));
-            
-        foreach (var mat in R_world2camArray) GC.KeepAlive(mat);
-        foreach (var mat in t_world2camArray) GC.KeepAlive(mat);
-        foreach (var mat in R_base2gripperArray) GC.KeepAlive(mat);
-        foreach (var mat in t_base2gripperArray) GC.KeepAlive(mat);
     }
 
     /// <summary>
@@ -3516,111 +1943,6 @@ static partial class Cv2
     }
 
     /// <summary>
-    /// filters off speckles (small regions of incorrectly computed disparity)
-    /// </summary>
-    /// <param name="img">The input 16-bit signed disparity image</param>
-    /// <param name="newVal">The disparity value used to paint-off the speckles</param>
-    /// <param name="maxSpeckleSize">The maximum speckle size to consider it a speckle. Larger blobs are not affected by the algorithm</param>
-    /// <param name="maxDiff">Maximum difference between neighbor disparity pixels to put them into the same blob. 
-    /// Note that since StereoBM, StereoSGBM and may be other algorithms return a fixed-point disparity map, where disparity values 
-    /// are multiplied by 16, this scale factor should be taken into account when specifying this parameter value.</param>
-    /// <param name="buf">The optional temporary buffer to avoid memory allocation within the function.</param>
-    public static void FilterSpeckles(InputOutputArray img, double newVal, int maxSpeckleSize, double maxDiff,
-        InputOutputArray? buf = null)
-    {
-        if (img is null)
-            throw new ArgumentNullException(nameof(img));
-        img.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.stereo_filterSpeckles(img.CvPtr, newVal, maxSpeckleSize, maxDiff, ToPtr(buf)));
-        GC.KeepAlive(img);
-        GC.KeepAlive(buf);
-        img.Fix();
-    }
-
-    /// <summary>
-    /// computes valid disparity ROI from the valid ROIs of the rectified images (that are returned by cv::stereoRectify())
-    /// </summary>
-    /// <param name="roi1"></param>
-    /// <param name="roi2"></param>
-    /// <param name="minDisparity"></param>
-    /// <param name="numberOfDisparities"></param>
-    /// <param name="SADWindowSize"></param>
-    /// <returns></returns>
-    public static Rect GetValidDisparityROI(Rect roi1, Rect roi2,
-        int minDisparity, int numberOfDisparities, int SADWindowSize)
-    {
-        NativeMethods.HandleException(
-            NativeMethods.stereo_getValidDisparityROI(
-                roi1, roi2, minDisparity, numberOfDisparities, SADWindowSize, out var ret));
-        return ret;
-    }
-
-    /// <summary>
-    /// validates disparity using the left-right check. The matrix "cost" should be computed by the stereo correspondence algorithm
-    /// </summary>
-    /// <param name="disparity"></param>
-    /// <param name="cost"></param>
-    /// <param name="minDisparity"></param>
-    /// <param name="numberOfDisparities"></param>
-    /// <param name="disp12MaxDisp"></param>
-    public static void ValidateDisparity(InputOutputArray disparity, InputArray cost,
-        int minDisparity, int numberOfDisparities, int disp12MaxDisp = 1)
-    {
-        if (disparity is null)
-            throw new ArgumentNullException(nameof(disparity));
-        if (cost is null)
-            throw new ArgumentNullException(nameof(cost));
-        disparity.ThrowIfNotReady();
-        cost.ThrowIfDisposed();
-
-        NativeMethods.HandleException(
-            NativeMethods.stereo_validateDisparity(
-                disparity.CvPtr, cost.CvPtr, minDisparity, numberOfDisparities, disp12MaxDisp));
-
-        disparity.Fix();
-        GC.KeepAlive(disparity);
-        GC.KeepAlive(cost);
-    }
-
-    /// <summary>
-    /// reprojects disparity image to 3D: (x,y,d)->(X,Y,Z) using the matrix Q returned by cv::stereoRectify
-    /// </summary>
-    /// <param name="disparity">Input single-channel 8-bit unsigned, 16-bit signed, 32-bit signed or 32-bit floating-point disparity image.</param>
-    /// <param name="_3dImage">Output 3-channel floating-point image of the same size as disparity. 
-    /// Each element of _3dImage(x,y) contains 3D coordinates of the point (x,y) computed from the disparity map.</param>
-    /// <param name="Q">4 x 4 perspective transformation matrix that can be obtained with stereoRectify().</param>
-    /// <param name="handleMissingValues">Indicates, whether the function should handle missing values (i.e. points where the disparity was not computed). 
-    /// If handleMissingValues=true, then pixels with the minimal disparity that corresponds to the outliers (see StereoBM::operator() ) are 
-    /// transformed to 3D points with a very large Z value (currently set to 10000).</param>
-    /// <param name="ddepth">he optional output array depth. If it is -1, the output image will have CV_32F depth. 
-    /// ddepth can also be set to CV_16S, CV_32S or CV_32F.</param>
-    public static void ReprojectImageTo3D(InputArray disparity,
-        OutputArray _3dImage, InputArray Q,
-        bool handleMissingValues = false, int ddepth = -1)
-    {
-        if (disparity is null)
-            throw new ArgumentNullException(nameof(disparity));
-        if (_3dImage is null)
-            throw new ArgumentNullException(nameof(_3dImage));
-        if (Q is null)
-            throw new ArgumentNullException(nameof(Q));
-        disparity.ThrowIfDisposed();
-        _3dImage.ThrowIfNotReady();
-        Q.ThrowIfDisposed();
-
-        NativeMethods.HandleException(
-            NativeMethods.stereo_reprojectImageTo3D(
-                disparity.CvPtr, _3dImage.CvPtr, Q.CvPtr, handleMissingValues ? 1 : 0, ddepth));
-
-        _3dImage.Fix();
-        GC.KeepAlive(disparity);
-        GC.KeepAlive(_3dImage);
-        GC.KeepAlive(Q);
-    }
-
-    /// <summary>
     /// Computes an optimal affine transformation between two 3D point sets.
     /// </summary>
     /// <param name="src">First input 3D point set.</param>
@@ -3887,139 +2209,6 @@ static partial class Cv2
         GC.KeepAlive(afterPoints);
         GC.KeepAlive(possibleSolutions);
         GC.KeepAlive(pointsMask);
-    }
-
-    /// <summary>
-    /// corrects lens distortion for the given camera matrix and distortion coefficients
-    /// </summary>
-    /// <param name="src">Input (distorted) image.</param>
-    /// <param name="dst">Output (corrected) image that has the same size and type as src .</param>
-    /// <param name="cameraMatrix"> Input camera matrix</param>
-    /// <param name="distCoeffs">Input vector of distortion coefficients (k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6]]) of 4, 5, 
-    /// or 8 elements. If the vector is null, the zero distortion coefficients are assumed.</param>
-    /// <param name="newCameraMatrix">Camera matrix of the distorted image. 
-    /// By default, it is the same as cameraMatrix but you may additionally scale 
-    /// and shift the result by using a different matrix.</param>
-    public static void Undistort(InputArray src, OutputArray dst,
-        InputArray cameraMatrix,
-        InputArray distCoeffs,
-        InputArray? newCameraMatrix = null)
-    {
-        if (src is null)
-            throw new ArgumentNullException(nameof(src));
-        if (dst is null)
-            throw new ArgumentNullException(nameof(dst));
-        if (cameraMatrix is null)
-            throw new ArgumentNullException(nameof(cameraMatrix));
-        src.ThrowIfDisposed();
-        dst.ThrowIfNotReady();
-        cameraMatrix.ThrowIfDisposed();
-
-        NativeMethods.HandleException(
-            NativeMethods.imgproc_undistort(src.CvPtr, dst.CvPtr, cameraMatrix.CvPtr,
-                ToPtr(distCoeffs), ToPtr(newCameraMatrix)));
-
-        GC.KeepAlive(src);
-        GC.KeepAlive(dst);
-        GC.KeepAlive(cameraMatrix);
-        GC.KeepAlive(distCoeffs);
-        GC.KeepAlive(newCameraMatrix);
-        dst.Fix();
-    }
-
-    /// <summary>
-    /// initializes maps for cv::remap() to correct lens distortion and optionally rectify the image
-    /// </summary>
-    /// <param name="cameraMatrix"></param>
-    /// <param name="distCoeffs"></param>
-    /// <param name="r"></param>
-    /// <param name="newCameraMatrix"></param>
-    /// <param name="size"></param>
-    /// <param name="m1Type"></param>
-    /// <param name="map1"></param>
-    /// <param name="map2"></param>
-    public static void InitUndistortRectifyMap(
-        InputArray cameraMatrix, InputArray distCoeffs,
-        InputArray r, InputArray newCameraMatrix,
-        Size size, MatType m1Type, OutputArray map1, OutputArray map2)
-    {
-        if (cameraMatrix is null)
-            throw new ArgumentNullException(nameof(cameraMatrix));
-        if (distCoeffs is null)
-            throw new ArgumentNullException(nameof(distCoeffs));
-        if (r is null)
-            throw new ArgumentNullException(nameof(r));
-        if (newCameraMatrix is null)
-            throw new ArgumentNullException(nameof(newCameraMatrix));
-        if (map1 is null)
-            throw new ArgumentNullException(nameof(map1));
-        if (map2 is null)
-            throw new ArgumentNullException(nameof(map2));
-        cameraMatrix.ThrowIfDisposed();
-        distCoeffs.ThrowIfDisposed();
-        r.ThrowIfDisposed();
-        newCameraMatrix.ThrowIfDisposed();
-        map1.ThrowIfNotReady();
-        map2.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.imgproc_initUndistortRectifyMap(
-                cameraMatrix.CvPtr, distCoeffs.CvPtr, r.CvPtr, newCameraMatrix.CvPtr, size, m1Type, map1.CvPtr, map2.CvPtr));
-
-        GC.KeepAlive(cameraMatrix);
-        GC.KeepAlive(distCoeffs);
-        GC.KeepAlive(r);
-        GC.KeepAlive(newCameraMatrix);
-        GC.KeepAlive(map1);
-        GC.KeepAlive(map2);
-        map1.Fix();
-        map2.Fix();
-    }
-
-    /// <summary>
-    /// initializes maps for cv::remap() for wide-angle
-    /// </summary>
-    /// <param name="cameraMatrix"></param>
-    /// <param name="distCoeffs"></param>
-    /// <param name="imageSize"></param>
-    /// <param name="destImageWidth"></param>
-    /// <param name="m1Type"></param>
-    /// <param name="map1"></param>
-    /// <param name="map2"></param>
-    /// <param name="projType"></param>
-    /// <param name="alpha"></param>
-    /// <returns></returns>
-    public static float InitWideAngleProjMap(
-        InputArray cameraMatrix, InputArray distCoeffs,
-        Size imageSize, int destImageWidth, MatType m1Type,
-        OutputArray map1, OutputArray map2,
-        ProjectionType projType, double alpha = 0)
-    {
-        if (cameraMatrix is null)
-            throw new ArgumentNullException(nameof(cameraMatrix));
-        if (distCoeffs is null)
-            throw new ArgumentNullException(nameof(distCoeffs));
-        if (map1 is null)
-            throw new ArgumentNullException(nameof(map1));
-        if (map2 is null)
-            throw new ArgumentNullException(nameof(map2));
-        cameraMatrix.ThrowIfDisposed();
-        distCoeffs.ThrowIfDisposed();
-        map1.ThrowIfNotReady();
-        map2.ThrowIfNotReady();
-
-        NativeMethods.HandleException(
-            NativeMethods.imgproc_initWideAngleProjMap(
-                cameraMatrix.CvPtr, distCoeffs.CvPtr, imageSize,
-                destImageWidth, m1Type, map1.CvPtr, map2.CvPtr, (int) projType, alpha, out var ret));
-
-        GC.KeepAlive(cameraMatrix);
-        GC.KeepAlive(distCoeffs);
-        GC.KeepAlive(map1);
-        GC.KeepAlive(map2);
-        map1.Fix();
-        map2.Fix();
-        return ret;
     }
 
     /// <summary>
