+++ src/OpenCvSharp/Internal/PInvoke/NativeMethods/imgproc/NativeMethods_imgproc_undistort.cs
@@ -1,0 +1,34 @@
+﻿using System.Runtime.InteropServices;
+
+#pragma warning disable 1591
+#pragma warning disable CA1401 // P/Invokes should not be visible
+#pragma warning disable IDE1006 // Naming style
+
+namespace OpenCvSharp.Internal;
+
+// ReSharper disable InconsistentNaming
+static partial class NativeMethods
+{
+    [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
+    public static extern ExceptionStatus imgproc_drawFrameAxes(
+        IntPtr image, IntPtr cameraMatrix, IntPtr distCoeffs,
+        IntPtr rvec, IntPtr tvec, float length, int thickness);
+
+    [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
+    public static extern ExceptionStatus imgproc_undistort(
+        IntPtr src, IntPtr dst,
+        IntPtr cameraMatrix, IntPtr distCoeffs, IntPtr newCameraMatrix);
+
+    [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
+    public static extern ExceptionStatus imgproc_initUndistortRectifyMap(
+        IntPtr cameraMatrix, IntPtr distCoeffs,
+        IntPtr R, IntPtr newCameraMatrix,
+        Size size, MatType m1type, IntPtr map1, IntPtr map2);
+
+    [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
+    public static extern ExceptionStatus imgproc_initWideAngleProjMap(
+        IntPtr cameraMatrix, IntPtr distCoeffs,
+        Size imageSize, int destImageWidth,
+        MatType m1type, IntPtr map1, IntPtr map2,
+        int projType, double alpha, out float returnValue);
+}
