+++ src/OpenCvSharp/Modules/calib/Enum/CalibrationFlags.cs
