+++ src/OpenCvSharp/Modules/calib/Enum/ChessboardFlags.cs
