+++ src/OpenCvSharp/Modules/calib/Enum/FindCirclesGridFlags.cs
