+++ src/OpenCvSharp/Modules/calib/Enum/HandEyeCalibrationMethod.cs
