+++ src/OpenCvSharp/Modules/calib/Enum/RobotWorldHandEyeCalibrationMethod.cs
