+++ src/OpenCvSharp/Modules/geometry/Enum/EssentialMatMethod.cs
