+++ src/OpenCvSharp/Modules/geometry/Enum/ProjectionType.cs
