+++ src/OpenCvSharp/Modules/geometry/Enum/SolvePnPFlags.cs
