+++ src/OpenCvSharp/Modules/stereo/Enum/StereoRectificationFlags.cs
