+++ src/OpenCvSharp/Modules/stereo/StereoBM.cs
