+++ src/OpenCvSharp/Modules/stereo/StereoMatcher.cs
