+++ src/OpenCvSharp/Modules/stereo/StereoSGBM.cs
