+++ src/tools/build-opencvsharp-minimal-arm64.sh
@@ -58,9 +58,12 @@ cmake .. \
   -DOpenCV_DIR="$INSTALL_PREFIX/lib/cmake/opencv4" \
   -DNO_CONTRIB=ON \
   -DNO_STITCHING=ON \
-  -DNO_CALIB3D=ON \
+  -DNO_GEOMETRY=ON \
+  -DNO_CALIB=ON \
+  -DNO_STEREO=ON \
+  -DNO_PTCLOUD=ON \
   -DNO_VIDEO=ON \
-  -DNO_FEATURES2D=ON \
+  -DNO_FEATURES=ON \
   -DNO_FLANN=ON \
   -DNO_DNN=ON \
   -DNO_ML=ON \
